Bug 1752351 - Improve the throttled capture mochitest. r=jib
Turned out the refresh driver was not at all throttled in the test. rAF was
throttled but these are not always in sync.

This patch changes the approach to causing throttling by being closer to a real
scenario -- put the tab in the background by opening another tab in the
foreground.

This also meant that the method by which we were measuring the captured frame
rate became moot (videoElement.mozPaintedFrames), as the frames are not
composited in the background, and the compositor is what reports painted frames.

This patch changes the frame counting approach to a local peer connection and
the outbound-rtp stats' framesEncoded member.

Differential Revision: https://phabricator.services.mozilla.com/D137290

UltraBlame original commit: 34237ea4f82f1948fb1f7976ba19e3590d96cc72