Bug 1752486 - Replace the win64 clang 2 stages toolchain with a 1 stage one. r=firefox-build-system-reviewers,andi
As of bug 1751824, the only remaining use of the 2-stages Windows clang
toolchain is for llvm-profdata in the generate-profile runs.

We don't need that to be built by clang itself (whereas it was desirable
to get compiler-rt from stage 2 before bug 1751824), so we can drop that
build to one stage only.

And because as of bug 1751839 we can start building stage 2 from a
separate stage 1 task, we do that for the Windows clang toolchain, like
we do for the Linux one.

Also, make the win64-clang-toolchain alias point to that stage 1 build,
like for Linux.

Differential Revision: https://phabricator.services.mozilla.com/D137257

UltraBlame original commit: 467d477bf6584bc119fa1e3b8b26d3a49e0b1293