Bug 1746750 - Part 1. Refactoring to support OffscreenCanvasRenderingContext2D. r=lsalzman
This patch is a non-functional change which allows the shutdown observer
to be overridden (since workers have a different mechanism). It also
fixes a static counter to be atomic, and adds mOffscreenCanvas to the
cycle collection tracking.

Differential Revision: https://phabricator.services.mozilla.com/D135352

UltraBlame original commit: b43727bd95d6a65784a1601b9cfc4ba3070e4e87