Bug 1750780 [wpt PR 32436] - Refactor srcdoc to use navigation params., a=testonly
Automatic update from web-platform-tests
Refactor srcdoc to use navigation params.

This CL removes the assumption that the owner of a srcdoc iframe lives
in the same process as the iframe's comment. It does this by sending
the srcdoc content to the browser in a mojo IPC, and sends it back
to the renderer via CommitNavigationParams. The content value is also
stored on the FrameTreeNode, and as such will be available in the
history.

Bug: 1233143
Change-Id: Ie511980dc6dfa4ec4cea18e4b1baf78c66a2a420
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3300866
Reviewed-by: Hiroki Nakagawa <nhirokichromium.org>
Reviewed-by: danakj chromium <danakjchromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzognichromium.org>
Reviewed-by: Alex Moshchuk <alexmoschromium.org>
Commit-Queue: James Maclean <wjmacleanchromium.org>
Cr-Commit-Position: refs/heads/main{#963191}

--

wpt-commits: 08176270bfb760dcf995908fccec46c7514a38b6
wpt-pr: 32436

UltraBlame original commit: c78de81ee4cac5bff360f656fdb564612cfc3821