Bug 1746104 - part 5-3: Make `SelectionChangeDataBase` treat no range case r=m_kato
Unfortunately, it cannot has `Maybe` because of used by a union.  Therefore,
it needs to manage with `bool` members of whether it's initialized with or
without error and whether it has a range.

Note that this changes `nsITextInputProcessorCallback.idl` which is currently
only for automated tests too.  The new behavior will be tested by the last
patch in this bug.

Differential Revision: https://phabricator.services.mozilla.com/D137424

UltraBlame original commit: fa7bd59d3ed663fa57c5ebf3510eb6c81f7a89f7