Bug 1754632 - Fix customization density menuitem focus style. r=Gijs
You can tab to focus a menuitem in the customize mode density menu,
and focusing one of the menuitems will preview the corresponding
density mode in the browser chrome. But the menuitem won't be styled
any differently, so it could be a bit confusing.

Differential Revision: https://phabricator.services.mozilla.com/D138387

UltraBlame original commit: ee5dc000ab462230332797c4768e75fc6c944e81