Bug 1753487 - Make `nsIContent::GetEditingHost()` not assume that itself is an element if there is no parent element r=peterv
If the instance is not an element and is a child of `Document` or something
non-element node, the method calls `this->AsElement()` in non-element instance.

It should check whether it's an element node or not when there is no editable
parent element.

Differential Revision: https://phabricator.services.mozilla.com/D138230

UltraBlame original commit: 80590d5a58300983cb9dc4688146d2dd8494beeb