Bug 1753718 - Pre: Prefix child process stdout lines with PID>; add stdout line callback. r=bytesized
We may get multiple lines or incomplete lines from the pipe, so we
need to split the data and keep the leftover.  This makes debugging a
little more pleasant and allows for the consumer to react to stdout as
it is read.

Differential Revision: https://phabricator.services.mozilla.com/D138222

UltraBlame original commit: 0f3d9fa6bec8915187a4bccaa3090a412ab9ed0f