Bug 1749298 - Part 3: Only trace finalization records if the registry's global is alive r=sfink
This adds a set of record object wrappers to the global data and traces them as
part of global tracing.

This may be cross compartment edges. They are not in the CCW map but are
handled by the extra sweep group edges added by the previous patch.

Differential Revision: https://phabricator.services.mozilla.com/D138438

UltraBlame original commit: ca41fc23ea6af6c2ef6b2615b2953665aff396b7