Bug 1726782 - Enable PGO on x64 and arm64 mac clangs. r=firefox-build-system-reviewers,andi
This should overcome the speed gap between Xcode clang and ours.

We use the profile for the linux build because gathering a profile on actual
macs would be too resource-intensive, and the result with the linux profile
is good enough.

Differential Revision: https://phabricator.services.mozilla.com/D138512

UltraBlame original commit: 17bfaad3db80d6400998309d447d67aaa2e1b971