Bug 1754323 - Ensure safe-area calculations aren't too off. r=m_kato
The fix here is the EnsureAtMost call, but we can simplify a bit of the
related code by replacing the four conditionals with an
EnsureAtLeast call too, and using Deflate() rather than the manual
calculations.

This is sort of a wall paper over bug 1754323, but should be safe and
prevents ridiculously big safe areas.

Differential Revision: https://phabricator.services.mozilla.com/D138453

UltraBlame original commit: 10d25b442074ae503a174a6d226bcb285fe531c4