Bug 1755021 - Don't clone and force OAs for partitioned principal if SharedWorker's principal is not a content princpial. r=dom-worker-reviewers,asuth
We don't need to clone and force OAs for partitioned principal if the
sharedWorker is using a null principal. It's because creating storage
for a null principal will fail anyway and try to clone and force OAs for
a null principal will also fail.

We should only do this for content principals.

Differential Revision: https://phabricator.services.mozilla.com/D138550

UltraBlame original commit: 1d277bf34ff7184c4e981ab1f7370f40f8279f90