Bug 1432343 - [devtools] Mutate messagesById in console redux state. r=ochameau.
The property is renamed, prefixed with `mutable` to better highlight that it is
special; the associated selector is renamed as well.
The only change necessary was adding a `messageCount` prop to the `ConsoleOutput`
component so its lifecycle method could check if the number of messages changed.

Differential Revision: https://phabricator.services.mozilla.com/D138293

UltraBlame original commit: 6714473d1152e4330ce639d8c86914ce3c72462d