Bug 1753196 - [devtools] RuleView: inherited properties marked as overridden incorrectly r=nchevobbe
When both the current and the earlier props for a given property name are inherited, the current prop should not be marked as overridden if the earlier prop was marked as !important.
!important should only come into play if the property where it was set is not inherited.
Regardless of this, properties are considered in descending order of specificity, so the earlier rule should take precedence.

Differential Revision: https://phabricator.services.mozilla.com/D138895

UltraBlame original commit: bc480c50be30183620d3a9ef29ad74458da84cdc