Bug 1755686 - Fix screen coordinate space confusions on PiP windows. r=mhowell
While at it, consistently use outerHeight to track window sizes. It
shouldn't matter but it's better to be consistent (we open the window
with `outerHeight=${desiredHeight}` and so on).

Note that for the window restore to work properly this needs bug 1755134.

Differential Revision: https://phabricator.services.mozilla.com/D138939

UltraBlame original commit: 01d8e901bb369cad52f885808fd0d8ebb5176ca7