Bug 1756162 - Make the visibility of the "Top Pick" checkbox depend entirely on whether the best match feature is enabled. r=preferences-reviewers,Gijs
We've made a few changes to this "Top pick" checkbox recently based on shifting
Product requirements, and the problem here is that the checkbox used to be
inside the Firefox Suggest container, but we recently moved it outside the
container. The Firefox Suggest container is properly hidden or shown depending
on whether the Suggest feature is enabled, so when the checkbox was inside the
container and Suggest was disabled, the checkbox properly got hidden too. Now
that the checkbox is outside that container, its visibility needs to entirely
depend on whether the best match feature is enabled.

So all this revision does is move the checkbox's `hidden` assignment from inside
the "is Suggest enabled" if-block to outside the if-block. It also sets
`hidden=true` on the checkbox in the markup for good measure.

I also improved the test so it checks every combination of enabled statuses
between the Suggest and best match features.

Depends on D138987

Differential Revision: https://phabricator.services.mozilla.com/D139161

UltraBlame original commit: 536fddf7da0517b43fcb92b27293bcf3ba06d889