Bug 1756727: Return E_NOTIMPL for a remote document in MsaaDocAccessible::get_accValue. r=morgan
The code previously assumed that the base MsaaAccessible implementation would return failure for a remote Accessible.
That is no longer true, so we crashed when we try to get the local DocAccessible.
For now, we explicitly check for remote and return E_NOTIMPL in that case.

Differential Revision: https://phabricator.services.mozilla.com/D139450

UltraBlame original commit: 25f40d9be64ea0530572a5cb4ddb8ad21bb3bbd6