Bug 1741899 - Detach session from GeckoView when setting it to other instance. r=calu
Historically, we wanted apps to be deliberate about setting/unsetting sessions
on a GeckoView instace because we were worried that they might leak a
GeckoSession (which would leak a tab running in Gecko), so we expected apps to
call `view.releaseSession` and collect the returned session.

We also expected apps's code to be more straightforward and clearly have a
`release`/`set` session pair that could be verifiably happening in order.

All of the above assumptions are broken in AC, where the code for
`releaseSession` and `setSession` is driven by a reactive data store.

The state management piece is also so hard for AC that they call
`releaseSession` conservatively in a lot of places and never actually look at
the return value.

To more align GeckoView with AC's architecture this patch moves into the
direction where apps only `setSession` and never `release`.

GeckoView will release the session from the old GeckoView when
appropriate.

The only problem with the above is that we might get into a state
where the app thinks a  GeckoView is displaying a session while really it isn't
because the session was set to a different GeckoView.

To detect the above case we introduce a new flag called `mIsSessionPoisoned`
which is set whenever the GeckoView instance used to hold a reference to a
session which has since gotten attached to a different window. If the app tries
to display such a GeckoView, we will throw an exception in `onAttachedToWindow`.

After this patch, apps won't need to manage `release`/`set` manually and if
they somehow end up in a broken state (e.g. because the app has a bug) we still
correctly crash.

This avoids crashing if the `release`/`set` pairs are out of order (e.g.
because `onAttachedToWindow`, `onDetachFromWindow` arrive out of order) which
is our leading theory for the crash in the bug attached to this patch.

Differential Revision: https://phabricator.services.mozilla.com/D139518

UltraBlame original commit: a7a1766b9d39de3d9716b4cb3e10fde899032487