Bug 1756663 - Add test for id option of history.onTitleChanged r=willdurand
In bug 1678611, the id property was added to this event, without unit
tests. The existing test didn't pick up the change because it checked
for a specific property. This patch adds a check of the whole event
object to get coverage for the id property AND to avoid this kind of
issue in the future.

Differential Revision: https://phabricator.services.mozilla.com/D139408

UltraBlame original commit: 2d6d35d5f6f7004a0436f60fd0d64f969ff6fb6b