Bug 1752445 [wpt PR 32578] - App history: implement transition.finished promise, a=testonly
Automatic update from web-platform-tests
App history: implement transition.finished promise

This also changes the abort reason used to signal abort on canceled navigations to be the same as the error passed to navigateerror and used for promise rejections. (This is tested by how the tests assert all errors are the same object.)

Also, in the course of working on this, I realized that NavigateReaction::_transition was always equal to AppHistory::_transition, so we could remove the former instead of checking them for equality. This led to a bit of refactoring.

This contains a large overhaul of tests related to appHistory.transition and further refinement of the previously-introduced tests related to ordering, to be more exhaustive and to use a small event-recorder framework. It consolidates all tests about appHistory.transition into the ordering tests (with the directory now renamed ordering-and-transition). This is because both types of tests were setting up various situations and checking either the ordering of events (often not exhaustively), or the properties of appHistory.transition, or both.

Bug: 1183545
Change-Id: Idc914a98a4dd6579d003806a992b710881944770
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3413934
Reviewed-by: Nate Chapin <japhetchromium.org>
Commit-Queue: Domenic Denicola <domenicchromium.org>
Cr-Commit-Position: refs/heads/main{#966216}

--

wpt-commits: 197bef9817160f643ea4bc50ce980562c2a4b61f
wpt-pr: 32578

UltraBlame original commit: 88003e1cdd3f2ec8b6cbce3f622b5fc68ff98e25