Bug 1752663 - Don't override tabpanel background unless forcing colors. r=dao
Blaming this code goes all the way through interesting history. This
code was originally made conditional in bug 1379587. At that time, the
use_system_colors pref wasn't true on Windows by default (it is now).

I think we only really want to do this if we're forcing content colors.
Otherwise we really want to match the new tab page colors, and the
use_system_colors pref doesn't change that.

The code in the condition is not perfect, mind you. In particular, it
should deal with document_color_use = 1 + high contrast theme, plus the
canvas color should react to the content color-scheme, not the browser
color-scheme. I thought of how to best fix that, and I think probably a
new system color for chrome to expose this makes sense, or alternatively
a ChromeUtils API or something, because otherwise reverse-engineering
everything that goes on in layout/style/PreferenceSheet.cpp is going to
be a huge mess.

But I'd rather do that in a follow-up bug, since it's nothing new and
this should fix Windows' default behavior.

Differential Revision: https://phabricator.services.mozilla.com/D137377

UltraBlame original commit: 8e78617fe28ca5b488a1fcaddb5e97d404f759cb