Bug 1739197 - Ensure listeners are connected before enqueueing events in TestPacer, r=bwc
Without this change, it was possible for the event dispatched in the
"Past" test case to be fully dispatched and notified on the background
thread pool before the call to WaitFor() connected a listener for the
event. This would lead to a test failure due to the observed
notification being the duplicated event fired 10 seconds after the
previous event was supposed to dispatch.

Differential Revision: https://phabricator.services.mozilla.com/D139750

UltraBlame original commit: 559391cb798ca9902a885a06ea07c921942f669f