Bug 1757488 - Disable best match blocking. r=daisuke
This puts the blocking UI behind a preference. It sounds like we may want to
enable blocking again later, so I don't want to remove all the code yet, and in
the meantime we should continue to run the related tests so we don't miss
regressions.

Note that this only puts the UI behind a pref. The API for blocking in the quick
suggest provider is not behind a pref since nothing will use it when the UI
affordances are removed.

This also modifies the test so it checks the key shortcut (Shift-Delete) when
blocking is both enabled and disabled.

Differential Revision: https://phabricator.services.mozilla.com/D139903

UltraBlame original commit: 6d6ced577f1eaf38d2c8f0a24edbd0a97cd89c6b