Bug 1755692 - Fix error in `observeActivity` `start` event listener.  r=rpl.
The event listener callback wasn't bound to this, so the access to `activityErrorMap`
was failing.
This patch slightly modifies the event listener so we use a fat arrow function,
as well as adding the `once` option, since we were removing the event listener
the first time the callback was called.

Differential Revision: https://phabricator.services.mozilla.com/D138894

UltraBlame original commit: dde39664e666e219fe31a7d8794c19d58d283354