Bug 1757911 - part 4: Make `HTMLEditor::DoSplitNode` notify `RangeUpdater` r=m_kato
Similar to the previous patch, in the splitting a node case,
`SplitNodeWithTransaction` notifies `RangeUpdater`.  Currently, this must not
cause any trouble, but for consistency with `DoJoinNodes`, `DoSplitNode` does
it instead.

Different from the previous patch, `SelAdjSplitNode` is not called if splitting
the node fails.  Therefore, it's enough to do it at end of `DoSplitNode`.

Differential Revision: https://phabricator.services.mozilla.com/D140464

UltraBlame original commit: 8fb082fb987ec440c73caf8675508fea0f5565d0