Bug 1758684 - Inherit modified flag from subtests when updating wpt metadata, r=karlcow
The wpt-update code ensures that it only writes data in cases the file
is changed by having a 'modified' flag on each metadata file. However
this wasn't being correctly set on the manifest in the case that only
a subtest was modified, and not a top level test.

It turns out that in the existing code the test modified flag is
usually set anyway, except in the case that the test itself has a
conditional status that doesn't require an update. Therefore this
wasn't caught by any of the tests, and hadn't caused lots of issues
with updates.

Differential Revision: https://phabricator.services.mozilla.com/D140713

UltraBlame original commit: 24c5f6b62ba23b7f1eadb18f42dae5a5700e4e03