Bug 1759107 - Don't add gecko profiler stack frames when unmarking gray things on helper threads r=jandem
This is a null pointer dereference because TlsContext is null on helper
threads.

We don't support adding profiler entries like this on helper threads at the
moment, but it's possible and we should maybe do this in the future.

Differential Revision: https://phabricator.services.mozilla.com/D140819

UltraBlame original commit: 5af2e0c7d94b02c9998b733effaed23079442940