Bug 1758624 - [devtools] Make Object previewers faster. r=bomsy.
- Accessing DebuggerObject.class is showing up in profile as it's a getter, so we
cache it before calling the previewers, and pass it as an extra argument.
- Remove object destructuring in some function signatures as it's
doing extra work we might end up not needing if the previewer isn't the one
that should be used for the object.
- For some previewers, add Sets for className to match so the lookup is faster
- Avoid checking `isWorker` upfront as it has some overhead

Test for Restricted grip is updated to check that we do get the url in preview.

Differential Revision: https://phabricator.services.mozilla.com/D140603

UltraBlame original commit: c4eb9bf9760ae15efba85024dac7943e6fed2022