Bug 1270746 part 1 - Remove hasInstance hook from CType objects. r=iain
We can instead rely on the default behavior for `instanceof` in `JS::OrdinaryHasInstance`.

This works because these objects are callable (since the JSClass has a call-hook) and
they have a `.prototype` property that returns the value in the proto slot.

Differential Revision: https://phabricator.services.mozilla.com/D141342

UltraBlame original commit: 910c5efb495b5d7693fc7221b184949c3bed55d7