Bug 1750987: Uninstaller support for Private Browsing enhancements. r=nalexander,mhowell
The most notable part of this is the rework of DeleteShortcuts to handle the new shortcuts logs in %PROGRAMDATA%.

Also of note is what we're not handling these new shortcuts in UpdateShortcutAppModelIDs/UpdateShortcutsBranding. As far as I can tell, these are one-off fixes to handle old shortcuts. Because we know that the new shortcut creation code will set from the start -- so there's no need to. This means that there's nothing to do during install or post-update.

Differential Revision: https://phabricator.services.mozilla.com/D139884

UltraBlame original commit: 93e272387c6abebd075af9bbc952b0995c644421