Bug 1760333 - Add CompilationInput::dump function. r=arai
This patch adds the ability to dump most of the content of CompilationInput.

To avoid adding extra dependencies in CompileOptions.h, a method named
`dumpWith` is added, and uses a template parameter to let the caller implement
what has to be done for each fields, with the expectation that C++ overloading
would be used to distinguish between the various types.

Differential Revision: https://phabricator.services.mozilla.com/D141511

UltraBlame original commit: f76d219e3ee378eddf05c56819b6db06af243ab0