Bug 1761111: Correct the removal of files in the vendor directory r=jewilde
The intention is to remove all the files in the vendoring directory
(which may be the yaml directory or may not be) except those
specified in 'keep'.

However convert_patterns_to_paths expects a list of patterns, and
we were providing it with only one. This resulted in it iterating
over '**' as a string (one iteration for '*' and a second for '*')

This prevented it from traversing recursively. Turning it into a
list fixes it.

Depends on D141901

Differential Revision: https://phabricator.services.mozilla.com/D141902

UltraBlame original commit: daed1e9ef335c4fa5d041db153bd88c826136cad