Bug 1748546 - Fix pre-existing perma-failures when running browserAction popup resize tests locally on Linux. r=mixedpuppy
browser_ext_browserAction_popup_resize.js was permafailing for me locally, even without the other changes part
of this issue, that was more annoying than usually because I wanted to be sure that the changes for the browserAction
persistent listeners were not breaking any of the existing tests but it was tricky without being able to see the
test to pass consistently also when running locally.

The reason behind the failure was a 1px difference in width for the browserAction popup.
Given that the tests in browser_ext_browserAction_popup_resize.js were already being tollerant
elsewhere about a 1px difference between the dimensions of the content compared to the dimensions
the test case was expecting, it seems reasonable to apply a similar tollerance in the
test helper used to wait for the content dimentions to reach a given set of expected dimentions
and make this test able to pass consistently when running locally.

Differential Revision: https://phabricator.services.mozilla.com/D141663

UltraBlame original commit: 48b01553f3cc2ece9b19a567fcf6a7178a1ff090