Bug 1748546 - Add missing browserActionFor definition in GeckoView ext-browserAction.js module. r=robwu,geckoview-reviewers,agi
The GeckoView ext-browserAction.js and ext-pageAction.js seems to not be fully providing
the browserActionFor/pageActionFor internal global helpers that their Desktop counterparts
are defining.

Technically this shouldn't be triggering any issue at the moment, because at the moment
it seems only used for the commands API (and also internally by the menu API apparently),
which are both not currently provided in GeckoView builds.

Anyway, it seems that they were still intented to be provided, because browserAction
does have it but incomplete version (whereas pageAction none at all).

It may be reasonable to move this in its own separate bugzilla issue, but in the meantime
I wanted to be sure to don't forget about it and so I'm attaching as a separate patch
in this set of changes for the two browserAction/pageAction APIs.

Depends on D141664

Differential Revision: https://phabricator.services.mozilla.com/D141666

UltraBlame original commit: ed616e2e0abb9f686f3d3ceab3059254f7ad2715