Bug 1761105 - Validate Nimbus features using variables instead of schemas r=emcminn
Not all features are complicated enough to require full on JSON schema to
describe their variables. For these simpler features, we generate (a lax) JSON
schema based on the contents of their `variables` field to validate them. This
way we can re-use the existing validation mechanism and do not have to write a
bespoke validator.

The schema generated is intended to be compatible with the one generated by
Experimenter.

Differential Revision: https://phabricator.services.mozilla.com/D141892

UltraBlame original commit: 9ddd244145bb2814e9c42ea59c3f6244623e6c33