Bug 1758729 - Ensure search engines loading different locales from WebExtension are correctly loaded when reloading engines. r=mcheang,jteow
The add-on manager may not have updated the default locale for the WebExtension, so we need to ensure we get the correct default locale.
Also correct the check for testing if the name is different to the existing engine.

Differential Revision: https://phabricator.services.mozilla.com/D141605

UltraBlame original commit: d6cd1fa1e6336c465819da94448d4b3aaf42cd3e