Bug 1759893 - Implement Parcelable for WebNotification. r=jonalmeida
Apps want to persist a WebNotification so that they can respond to taps on it
when the app is not alive anymore.

To achive this, we make WebNotification serializable through Parcel so that the
app can store it somewhere.

Differential Revision: https://phabricator.services.mozilla.com/D141406

UltraBlame original commit: 3d305df975f39e0474fce4a7d34cfa2a8d5c604e