Bug 1761437 - Ensure the shared font namespace is allocated by the client. r=gw
The shared font namespace was getting allocated by the RenderBackend, while
we otherwise allocated namespaces for documents on the client. We need to make
sure that if namespace allocation happens on the client, that it also happens
on the client for the shared font namespace. This adds to RendererOptions to
make passing in a shared font namespace allocated by the client possible.

Differential Revision: https://phabricator.services.mozilla.com/D142381

UltraBlame original commit: 806e9fc2d511f085936058eb6fb6173ec0b6d364