Bug 1756592: Cache forced-return in ExceptionBailoutInfo r=tcampbell
My first draft of this patch cached cx->status, but that's a private member that isn't exposed by an accessor. Under the assumption that this was a deliberate choice, this version of the patch only caches whether we are doing a forced return.

Differential Revision: https://phabricator.services.mozilla.com/D140283

UltraBlame original commit: 40c4a9994215625e495b7690dc24e589187cf111