Bug 1761395 - part 4: Rewrite `HTMLEditor::SplitMailCiteElements` r=m_kato
This patch renames it to `HandleInsertParagraphInMailCiteElement` to explain
what it does clearer.  And make it takes most-distant ancestor mail-cite
element to avoid the situation that it does nothing.

Then, this is rewritten to just return new caret position where is new
`<br>` element which is inserted at the split point.

Then, only the last block is changed to refer the split node result to
avoid deleting same node in the case of new split direction.

Differential Revision: https://phabricator.services.mozilla.com/D142415

UltraBlame original commit: 62f2292245963158b3705ed1e26cbb958a627132