Bug 1763405 - Use a higher priority for the renderer thread. r=gfx-reviewers,jnicol
The renderer thread is at the end of the rendering pipeline. For smoothness it is important that it completes its work quickly as soon as it receives some. It does not produce work for other threads so we don't risk of causing some build up. On the other hand, the higher priority will help with with avoiding some cases of congestion of work coming from the content process.

This only bumps the priority of the renderer thread as opposed to the entire GPU process (Blink does the latter). This is because we currently have too many thread on the GPU process so there is a risk of completely starving other processes. It's still something we want to consider and might be safer after we have done the work of reducing the number of threads.

Differential Revision: https://phabricator.services.mozilla.com/D143056

UltraBlame original commit: fdfe49e89df7e569ea51228427a290668b5d5f6c