Bug 1763574 - Fix non-texture dirty rects incorrectly adding render tasks r=gfx-reviewers,lsalzman
If the overall surface dirty rect is large enough to include prims
on texture surface tasks, but those texture surface tasks are not
dirty, we could accidentally consider a primitive visible and needing
to be rendered. If that prim has a clip mask render task, but there
are no invalid picture cache tiles, there are no parent tasks for
the clip mask tasks to be attached to, which means the render task
graph never determines a parent pass to free those tasks on.

Differential Revision: https://phabricator.services.mozilla.com/D143227

UltraBlame original commit: 8018c7481e1ef3887c7481bcf10ccb85682543d9