Bug 1761932: Call recommendation sources in parallel and apply source-specific weights. r=Standard8
There are some non-final pieces here that will be improved in the next patch but
this makes most of the changes to how we generate and combine recommendations.

It does not yet sum the source specific scores instead still using the highest
so the same scores show up in tests. The next patch will change the calculation.

Differential Revision: https://phabricator.services.mozilla.com/D142792

UltraBlame original commit: 26184f0ae876de6042a292850bdfe5d1bcf52c54