Bug 1762402 [wpt PR 33449] - Pre-paint: Don't miss AREA elements in culled inlines., a=testonly
Automatic update from web-platform-tests
Pre-paint: Don't miss AREA elements in culled inlines.

We'd miss them when block-fragmented. We used to assume that the parent
of the AREA would occur in the last block fragment, but that's not a
safe assumption to make.

Instead, enter it when visiting the parent for the first time. We only
have a flag for the first time visiting, not for the last time, but it
doesn't really matter, as long as we visit it once, and only once.

Bug: 1308078
Change-Id: If549ad14b81e683af47b08c13fe42747c853e5a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3562999
Reviewed-by: Koji Ishii <kojiichromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhuchromium.org>
Commit-Queue: Morten Stenshorne <mstenshochromium.org>
Cr-Commit-Position: refs/heads/main{#987805}

--

wpt-commits: b451545dffd4dec0ce4652f81658abdece08872a
wpt-pr: 33449

UltraBlame original commit: 45f20e1a93e03a2ddeb8a6609acd4f5e7c38d6c6