Bug 1761207 - Calculate the expected expired days directly from the pref in the test browser_contentBlockingTelemetry.js. r=anti-tracking-reviewers,pbz
In the test, we check if record to the correct bucket for the remaining
expired time of the storage access permission. The expected values were
using static value, which will have problem if we change the expired
time via the prefs.

We change to calculate the expected expired time directly from the pref
so this test can work even we change the pref value.

Differential Revision: https://phabricator.services.mozilla.com/D143475

UltraBlame original commit: af56f3b0e69515e8d82c2d4d031ebe13c8507a2c