Bug 1764564 - Remove the code to ignore the Windows SDK directory from coverage. r=marco
It turns out the only reason it was excluded before is that the SDK used
to be downloaded into the source directory. Since it has moved out of
there, as per the code comment, we no longer need to ignore the SDK and
the code can be removed.

Differential Revision: https://phabricator.services.mozilla.com/D143781

UltraBlame original commit: 709b507ed708763f0c809fe86e48bfcea071b7b6