Bug 1752332: Sanitize preferences in the shared memory region r=KrisWright
Before we would stick all prefs into an immutable shared memory
region and clear our HashMap.

Now we will stick all non-sanitized prefs into the immutable
shared memory region, save the sanitized prefs in a list
temporarily, clear the hashmap, and then repopulate the hashmap
with the sanitized prefs.

As a bit of underlying complexity, to do this we must do some
tricks with the Pref Name Arena which is a chunk of memory
dedicated to storing pref names. That goes away (and we want to
wipe it to save space) - so we just need to move the sanitized
pref names to a new arena.

Depends on D141421

Differential Revision: https://phabricator.services.mozilla.com/D141422

UltraBlame original commit: 10ac6886fa5e32e67c6691c655c0e963b48edb9d