Bug 1680243 - move memory bounds check trap out-of-line. r=jseward
When spectre mitigations are off there's a big win to be had from
moving bounds check traps out-of-line, as conditional forward branches
are predicted not-taken on x86/x64 (documented) as well as arm/arm64
(less well documented, but the Internet believes this).  So implement
this.

When mitigations are enabled it's not a clear-cut win, since we end up
executing the speculation fence and that is not free.  So don't branch
out-of-line to the trap then; instead use the existing bounds check
logic that falls through into the trap and branches across it for a
successful check.

Currently we're running with spectre mitigations enabled, but the main
beneficiary here is going to be memory64, and by the time memory64
ships it's likely we'll be able to disable spectre mitigations, at
least in COOP+COEP processes.  We should land the present patch now so
that further work on memory64 bounds checking can start with the best
possible code.

Drive-by fix: the out-of-line interrupt trap is rebranded as a generic
'resumable' trap, to better fit the implementation of the the
out-of-line bounds check trap as an 'aborting' trap in this patch.

Differential Revision: https://phabricator.services.mozilla.com/D142562

UltraBlame original commit: 5d7dc3032f97ef4a596235db2feb3dc8ab8bc59f