Bug 1763576 - [devtools] Remove unnecessary await on synchronous dispatches. r=bomsy
When dispatching a JS Object, this won't be dispatching an async action.
Dropping the await will better highlight the synchronous dispatch calls.
Note that when we do
```
dispatch({
  [PROMISE]: ...,
  ...
});
```
The promise middleware kicks in and the action becomes async.
So await is still relevant in front of these dispatches.

Differential Revision: https://phabricator.services.mozilla.com/D143143

UltraBlame original commit: c95859201259666efcd49533fe7cca1ddea5ab24