Bug 1765922 - Ensure/ReleaseBufferForMainThreadAddMarker from main thread in profiler_start/stop - r=florian
In child processes, Gecko Profiler's profiler_start/stop are running OFF the main thread, so the calls to Ensure/ReleaseBufferForMainThreadAddMarker should be dispatched to the main thread for them to actually create/destroy the main-thread buffer.

Note that this is not done in the Base Profiler, because it should always run on the main thread, and if it wasn't there would be no way to dispatch a task to the main thread -- and it wouldn't be too much of an issue anyway, there are few markers running outside of Gecko Profiler sessions.

Differential Revision: https://phabricator.services.mozilla.com/D144387

UltraBlame original commit: a5ff969a3d72c0bfba299e2637d3491f5112b4af