Bug 1763781 Part 2: Update test expectations. r=jgilbert
This gives us better correctness for 10-bit vp9 reftests. In the WPTs, we
turn off some tests on macos because of the difficulty that our
screenshots have in analyzing srgb image data. It seems to consistently
report 191,255,191,73 when we are expecting 255,0,0,255.

Differential Revision: https://phabricator.services.mozilla.com/D144228

UltraBlame original commit: 3b1422986294d2f38d6dee1e44635d38f015f793