Bug 1765912 - Allow different strides in nsDOMWindowUtils::CompareCanvases. r=aosmond
Currently nsDOMWindowUtils::CompareCanvases assumes that if the data from two different
canvases is the same, then the strides of the snapshots should be the same. However, this
is not the case for DrawTargetWebgl, and this can cause various reftests to fail when
accelerated Canvas2D is enabled. All we actually care about is that the size and data
contents is the same, whereas the stride has no consequence other than that we need to
note it for proper traversal of the data for both snapshots.

Depends on D144381

Differential Revision: https://phabricator.services.mozilla.com/D144382

UltraBlame original commit: 5b9814ccc8ea97336b881f28f6cfcc25da8ef50d