Bug 1735565 - disabling erroneous https-only test case to isolate error r=ckerschb
This iterates on the known-buggy https-only test, to ensure that
we are sure that this is the only erroneous test case.

Admittedly, this whole test should just not use documents/iframes for
tests, but for now, this here will do.

Differential Revision: https://phabricator.services.mozilla.com/D144797

UltraBlame original commit: 47c396db5affab94ffbbae74daa9d94ba440bee6