Bug 1767165 - Fix dragend coordinates in child processes after bug 1756241. r=tnikkel
Bug 1756241 changed nsDragService::mDragEndPoint to be relative to the
window. I didn't realize, but that point is passed as-is to child
widgets. Changing the widget event to actually have a widget made the
point be interpreted as widget-relative (which is correct in the parent
process, but not in the child process). Account for this and fix the
mDragEndPoint documentation while at it.

Differential Revision: https://phabricator.services.mozilla.com/D145182

UltraBlame original commit: 3a5b69d067b9ce5dfb18ee216765279324ddc783