Bug 1767087 - Part 6: Perform [[AlreadyResolved]] check in default resolving function case in NewPromiseReactionJob. r=yulia
RunFulfillFunction and RunRejectFunction are for NewPromiseReactionJob spec
section, but they're called also from some other cases to reuse the code.
That needs to be addressed in separate bug to make it clearer the relation.

Testcase is added in Part 7.

Depends on D145205

Differential Revision: https://phabricator.services.mozilla.com/D145206

UltraBlame original commit: c1d66dbdbdc20aaa45f66d514b021acfa5569638