Bug 1677559 - Trigger a clean rebuild for browsertime upgrades/installations. r=perftest-reviewers,kshampur
This patch causes us to trigger a rebuild for the python environment on browsertime upgrades and installations. This is needed for now because we are moving to python-based dependencies for visual-metrics processing and that requires a rebuild on the first upgrade/installation.

See bug 1766112 for the removal of this code.

Depends on D144464

Differential Revision: https://phabricator.services.mozilla.com/D144565

UltraBlame original commit: 7bbdcfc5d92dd30fa5eb7b603643c6cfb1d4fa6f