Bug 1767069 part 1: Standardize on "float-avoiding block" in var/function names to refer to block-level elements that cannot intersect floats. r=emilio
This patch does not impact behavior; it's just a handful of renamings and
documentation-updates.

Before this patch, our float-handling code has a bunch of variables and
functions with "replaced" in the name. But in fact these aren't necessarily
replaced; they're just block-level frames that elicit a "false" return-value
from "BlockCanIntersectFloats".  This category *includes* replaced boxes, and
it also includes frames that establish a formatting context, e.g.:
scrollframes, flex/grid containers, flow-root, table.

This patch seeks to clarify things by renaming these to refer to the fact that
they "avoid" floats, e.g. s/replacedBlock/floatAvoidingBlock/.

I've included a comment to reference this term in the BlockCanIntersectFloats()
documentation (note that the "can intersect" concept is the opposite of
"avoids").  I've also renamed a few "aFrame" variables to
"aFloatAvoidingBlock", for a few functions that are explicitly dealing with
this scenario and only expect to be passed frames that are in this category.

Differential Revision: https://phabricator.services.mozilla.com/D145131

UltraBlame original commit: 7dd612a44473e778fb818752d8f675ae843ce1cd