Bug 1766730: Finish non-local jumps in finally block r=jandem
This is the patch that everything has been building towards.

Instead of pushing a resume index and using retsub, we push a continuation index, which is used as the discriminant in a tableswitch. Index 0 is reserved for the fallthrough case. Other continuations store a kind (break, continue, or return) and a target control (the continued / broken loop; nullptr for returns). Because of nesting, the target is guaranteed to outlive the finally control.

We have to be a little careful in `emitNonLocalJump` to skip past the switch and finally that we're currently emitting.

Differential Revision: https://phabricator.services.mozilla.com/D145033

UltraBlame original commit: 0db6f17a0d7324524837aef6cc2389cbb66ff6ef