Bug 1759686 - patch 4 - Extend ServoCSSParser::ParseFontShorthandForMatching to (optionally) return font size in addition to family/style. r=emilio
When running in a Worker, the canvas2d SetFontInternal method can't call GetFontStyleForServo
because it doesn't have a canvasElement or presShell to pass. But we can use the method
ServoCSSParser::ParseFontShorthandForMatching to parse the canvas 'font' property, if we
extend it so as to also return the size from the font shorthand.

(This is incomplete by itself; Emilio's following patch fixes it up, so the intention is to land them together.)

Depends on D144186

Differential Revision: https://phabricator.services.mozilla.com/D144187

UltraBlame original commit: 8a2c9882e141e8e81594fdd6af6306ec44016604