Bug 1767395 disallow loading http/https scripts for systemprincipal r=ckerschb,tjr
This copies over the behavior for style & subdocument restrictions.
Admittedly, with this if/else spagetthi, it would be preferable to
turn this into restriction levels or lump some of the known-to-be-safe
prefs together, but I would prefer we wait a couple of cycles to
make sure this makes it all the way to release before we refactor.

Differential Revision: https://phabricator.services.mozilla.com/D145306

UltraBlame original commit: 2f9156dc45d4055cfd31b7488f2f63f0bb82e220