Bug 1767452 - Fix _getActiveGMPPlugins. r=Gijs,extension-reviewers,kmag
The deferred path is currently broken because we stopped notifying
'plugin-blocklist-loaded' when we removed NPAPI blocklisting in
6e00aba7e14f

The current conditions for taking the deferred path are also no longer needed.
Services.blocklist.isLoaded is always true since we removed the XML blocklist
backend in 06ca3c111fc7 so we don't need to check for it anymore. The atStartup
parameter was added in 1fa2a1d2d41e to avoiding checking isLoaded during
startup which we don't need anymore because we aren't checking isLoaded.

However, bug 1522930 delayed the start of the GMP provider so we still need
a deferred path. To make that work we add a notification when the GMP provider
starts and a way to check if it's been registered.

This will fix the GMP addons data getting stuck in the dummy state.

Differential Revision: https://phabricator.services.mozilla.com/D145329

UltraBlame original commit: f53a3900cbe50c2dce84bde8e903256728f418ba