Bug 1470781 - [devtools] Make check in browser_styleeditor_private_perwindowpb.js more specific. r=jdescottes.
For some reason, the new test added in the previous patch has an impact on this test:
when checking the disk cache, there's a `mochi.test:8888/tests.json` entry, and
since the test was only checking if there was an entry with the `mochi.test:8888` entry,
it was failing.
I wasn't able to understand why we have this entry in the cache, but it reveals
that the test check is a bit too broad, since what we cares about is simply to
know if the fetched stylesheet was added in the disk cache.
So here we're doing just that, checking that the fetched stylesheet wasn't added
to the disk cache.

Depends on D144723

Differential Revision: https://phabricator.services.mozilla.com/D145344

UltraBlame original commit: 1c7305c88a56222410bd68cb1f9ac2ab32a7138d