Bug 1763645 - wasm: Split out stack map header to prepare for serialization. r=jseward
This commit splits out the header fields of the stack map to
simplify serialization. The header is POD that can be serialized
separately from the trailing bitmap.

An interface for writing to a bitmap is added for deserialization.
An interface to finish collecting stackmaps without sorting is added.
The const qualification of nextInsAddr is modified.

Depends on D143400

Differential Revision: https://phabricator.services.mozilla.com/D143401

UltraBlame original commit: a1736cc48e510508e44e82b1ab8488c7e3ddff52