Bug 1767626 - Accept a 5 pixel error when comparing duplicate frames. r=perftest-reviewers,AlexandruIonescu
This patch updates browsertime to include a patch that accepts a 5 pixel error (instead of 0) when comparing the frames at the end of the recording to remove duplicate ones. With this change, we should no longer over-estimate the LastVisualChange.

Differential Revision: https://phabricator.services.mozilla.com/D145457

UltraBlame original commit: 8c8dc11475fa15885fd6fc35c428bf221a5a286f