Bug 1763606 - Part 3: Port TypedArray RadixSort function to C++. r=tcampbell
An almost direct translation of the JS implementation to C++. In contrast to
the JS implementation, we're using a smaller fallback to `std::sort`, because
the C++ implementation is faster, but we're also calling into `std::sort` for
large arrays to avoid allocating a large heap buffer.

As a further improvement, the radix sort implementation calls the counting sort
function for 16-bit types when the number of elements reaches `2^16`. That way
we can further reduce the number of allocated heap memory.

The pre- and post-process steps from the JS implementation can be omitted,
because we can directly interpret the TypedArray memory contents as unsigned
integers in C++.

The concurrent write access handling in `SortByColumn` is necessary to avoid
an out-of-bounds access when a different thread writes to the same shared
memory location. This was also an issue in the JS implementation, except the
only visible outcome in that implementation was that the internal `aux` array
could leak to the user.

Depends on D143286

Differential Revision: https://phabricator.services.mozilla.com/D143287

UltraBlame original commit: 403a6eebf0a4e26010e1b4b896fb0209b53f74ef