Bug 1768215 - Use the right referrer info for constructable stylesheets. r=jfkthame
Document::GetReferrerInfo is not what we want (that is the referrer of
the document itself).

What we want is to use the document _as_ the referrer. That's what
regular stylesheets do, and that fixes our chrome only rules enabled
check.

Differential Revision: https://phabricator.services.mozilla.com/D145794

UltraBlame original commit: 22be3d5cdf4a35da243c3a4543a0632cedf22fe0