Bug 1432901 - Part 1: Add module APIs to get the module object for a script and the namespace object for a module r=yulia
It's a bit annoying that we have to convert between the module record object
and module script like this. I filed bug 1766810 to change the APIs but it's a
resonable amount of work at this point.

This adds JS APIs to get the module object for a module script and get a module
namespace object. The former required some refactoring of JSScript::module
(which could return null) into isModule() and module().

Differential Revision: https://phabricator.services.mozilla.com/D145554

UltraBlame original commit: 92d99a38d1c5bc7a739529c29b3a07bc33387239