Bug 1768237 - Don't take the lock in gfxFontFamily::HasOtherFamilyNames before calling ReadOtherFamilyNames, because it does the locking internally. r=lsalzman
This was a trivial error, and accounts for some of the crashes listed. It's wrong for HasOtherFamilyNames
to lock the gfxFontFamily record, because it then calls ReadOtherFamilyNames which does locking internally.
Other callsites of ReadOtherFamilyNames don't lock before calling it, we shouldn't do so here either.

Differential Revision: https://phabricator.services.mozilla.com/D145970

UltraBlame original commit: 26992be2722220b529706d2a81019137440014aa