Bug 1760928 - Avoid slow WebGL texture initialization for shared cache pages. r=aosmond
WebGL tries to zero initialize new textures that are only partially filled with data so
that the previous content of video memory does not leak into the texture from outside
the canvas. However, this initialization can incur high overhead in profiles.

It is more advantageous to keep around a PBO with zero data that is already cached
in video memory, so that it can then be copied to the newly allocated texture to
allocate it. WebGL will notice that this copy overwrites the entire texture and so
will avoid its own internal initialization, giving an overall performance win.

Differential Revision: https://phabricator.services.mozilla.com/D145500

UltraBlame original commit: d414d08b34867bf79f462f08d5bd65c42aff6eda