Bug 1766572: Part 2 - Synchronize KillClearOnShutdown phases with AdvanceShutdownPhase. r=xpcom-reviewers,dthayer,nika
The concept of a shutdown phase is meant to be a finite and named period of time where:

1. we can query if we are in or beyond a given phase
2. we may want to do a fast shutdown
3. we inform the terminator that we reached the next phase
4. if needed/wanted, we notify observers of the associated topic
5. we KillClearOnShutdown smart pointers associated with this phase
6. we do any further, phase-individual cleanup until the next phase starts

`AdvanceShutdownPhase(WithoutNotify)` now provides the functionality for 1. through 5. and must be the only place that calls `KillClearOnShutdown` in order to avoid misalignments.
In doing so, it becomes also the only caller of `MaybeFastShutdown`.

Please refer to bug 1768581 as working on this patch made us think if we should re-work the order inside AdvanceShutdownPhase.

Differential Revision: https://phabricator.services.mozilla.com/D145083

UltraBlame original commit: 77c9472e4932f7070dfb008a2c4d0c174e307d0c