Bug 1726388 - Make it possible to not anti-alias at the bounds of the primitive. r=gfx-reviewers,lsalzman
The Antialiasing code sets up the bounds of a rect to anti-alias in the vertex shader for the fragment shader to compute the AA via a distance approximation.
The modified mix call selects between the edges of the current instance (could be a primitive segment) and the edges of the primitive based on the anti-aliasing edge flags. If an edge does not have the AA flag, then the aa will be computed based on the distance to the primitive's edge. This usually works because we tend to want AA around the primitive but not along the internal edges of potential segments of a primitive. However it also means that if we take the anti-aliasing code path (for example under a non-axis-aligned transform) then the edges of the primitive are always anti-aliased. In some cases we want to disable the anti-aliasing along the primtive, though, for example when something is split into multiple primitive like the linear gradient optimization.

This patch changes this mix call so that it selects between the segment bounds and some very large rect. If the primitive is not segmented, then the segment bounds are the bounds of the primitive. The effect is that AA is only applied to edges that have the flag set rather than segment edges with the flag set as well as any edge at the border of the primitive.

Differential Revision: https://phabricator.services.mozilla.com/D145935

UltraBlame original commit: 1d42c0e759baaa6fd65235de1aed56462e47d872