Bug 1755094 - Implement ChildActorDispatcher. r=calu
The existing EventDispatcher messaging infrastracture uses cpmm (content
process message manager) which is deprecated and will be removed for Fission.

The current behavior with fission is that cpmm is available only on in-frame
frames, so on any oop frame we're not able to send or receive eventDispatcher
messages.

To allow OOP frames actors to send messages, this patch implements a new type
of dispatcher, called ChildActorDispatcher which routes messages through the
actor messaging system.

We then catch these messages on the parent actor side and send them to Java.

To be able to reroute messages from the Parent actor we need to make sure that
each actor that we define extends GeckoViewActorParent, so we define an actor
parent for each missing child actor.

In the future we could consider moving GeckoViewActorParent to a native
implementation for performance reasons.

Differential Revision: https://phabricator.services.mozilla.com/D139073

UltraBlame original commit: fe2864aaa7acf9adb506fc9c61710707be9282e1