Bug 1760637 - Upgrade initPanel in instantEditBookmark.js to async r=mak
In addition to adding async/await, I also added early return statements
when we know initialization is occuring again, or if there is evidence
that it has been uninitialized.

A concrete example of this would be from the Library, where clicking on
on a folder/category on the left will run initPanel at the same time as
the highlighted folder/bookmark on the right changes (also runs initPanel).

Differential Revision: https://phabricator.services.mozilla.com/D141800

UltraBlame original commit: 4e81d38260422d3b1fa04e9f940a0c03ab7c44a3