Bug 1766087 - Fix test_blocklistchange.js to use correct versions r=Gijs
The "succesful" update part of the test referenced a XPI where the
version differed from the update manifest. This is not supported,
the update manifest's version should match the version from the XPI.

As noted in the bug, addon_update1.json and addon_update2.json also
specify incorrect versions, but as the XPIs are not actually fetched
that doesn't matter in the test.
(I will address that in the next patch.)

Differential Revision: https://phabricator.services.mozilla.com/D144496

UltraBlame original commit: d17f69e5fda2ca977101e134bde8886887530d99