Bug 1768660: Skip fewer values in buildExpressionStack r=jandem
For the innermost frame of debugger mode bailouts, the current approach effectively uses an allow-list to decide which stack slots need to be recovered. This is fragile to any future bytecode changes that keep values alive on the expression stack. It's also unnecessary: if we just recover slots that are included in the snapshot, and skip slots that don't have allocations, everything works out.

Differential Revision: https://phabricator.services.mozilla.com/D146525

UltraBlame original commit: d86140d2fd38138052eba5164efa5e1ec6c3fd0f