Bug 1766237 - [devtools] Get the WebExtensionPolicy from the ExtensionProcessScript global r=jdescottes,rpl,ochameau
The issue here is it seems like WebExtensionPolicy is not always available on the Cu global,
therefore failing when trying to get the extension policy name.
in this example when adding sources from the browser toolbox (i'm still not completely sure why)

The solution here is changing to using the ExtensionProcessScript which works for the web and browser
toolbox.
Also added a `console.warn` so we get some notification when this fails.

Differential Revision: https://phabricator.services.mozilla.com/D144787

UltraBlame original commit: 597785337a95b29ebbc11ed3c1834653e2552acd