Bug 1768734 - Make sure JS holders are cleared before a shutdown GC r=mccr8
Previously this was optimized to trace only gray roots in collecting zones, but
there was a bug in that it also skipped clearing roots because this happens
before the start of GC when zones are selected for collection.

Instead, make sure we only perform this optimziation when marking.

Differential Revision: https://phabricator.services.mozilla.com/D146585

UltraBlame original commit: 0b0b6c707f5b3730a58c57d99dcd4d6808360217