Bug 1769973 - [devtools] Localize Device headers in Settings modal. r=bomsy.
We already have translation strings in `device.properties`, so we can use them.
String for `custom` type was missing, so we add it.
Assertions are added in test to check that we do have the expected labels in the UI.

Differential Revision: https://phabricator.services.mozilla.com/D146784

UltraBlame original commit: e5e9bb9e93360c1c23504a004fdf309fcc361044