Bug 1762874. Make sure to draw swipe to nav ui element opaque if velocity pushes us over success threshold. r=hiro
Both the velocity and displacement contribute to the success but we only pass the displacement in the swipe events which browser-gestureSupport.js uses to draw the opacity of the ui element. Usually that is what we want, but if the velocity is large we might draw the ui element only very faintly but still navigate the page. We should make sure to draw the ui element so it's well visible.

Differential Revision: https://phabricator.services.mozilla.com/D142819

UltraBlame original commit: d6975c4a6b6fd6dca62f0e5f7f53601b06aafe6a