Bug 1770273: Annotate reftest 1673510-1.html as reliably failing on win7 (instead of skipped). r=jfkthame
Based on the retriggers on Try, the test seems to always fail on this platform.
So let's mark it as failing instead of skipped, so that we find out if this
issue becomes fixed at some point.

Differential Revision: https://phabricator.services.mozilla.com/D147554

UltraBlame original commit: d270cb82e282e396364f7e892a0f5f433dba3a68