Bug 1764179 - [devtools] Ensure that TargetFront is fully initialized when processing resources sent via the Watcher. r=jdescottes
When resources related to a given Target are notified via the Watcher actor,
we have to ensure waiting for the TargetFront to be fully initialized.
That's because code watching for resources expect the target front to be attached
and have its thread front already available.

Differential Revision: https://phabricator.services.mozilla.com/D143408

UltraBlame original commit: 9566586e3ec5fc70f97c4d6bc88dcf9a2f839e41