Bug 1770922 part 1 - Use a Label instead of CodeLabel in EnterJit trampoline. r=iain
Now that it's a plain `Label`, we don't need the `scratch` register for the jump
anymore, so we no longer have to worry about it conflicting with `JSReturnOperand`.

This frees up two registers on x86, one of which will be used in the next patch.

Also tidies up the register set code a little.

Differential Revision: https://phabricator.services.mozilla.com/D147261

UltraBlame original commit: b56f906e62edfb8082ede9bf984838a3c9f61d22