Bug 1771565, part 2 - Remove XUL stuff from test_bug564863.xhtml. r=smaug
This removes XUL stuff from the Mochitest plain, so we don't need
to rely on allowXULXBL.

It also fixes up test_bug564863-2.xhtml to run as a mochitest-chrome.
It removes all of the non-XUL stuff, as that's being tested in the existing
test. I also commented out the is(xul_cs.color, "rgb(0, 0, 0)") check
as that seems to fail when this test is run as a Mochitest chrome
(for everything, not just XUL) and I didn't want to investigate
why.

Differential Revision: https://phabricator.services.mozilla.com/D147558

UltraBlame original commit: 4957cbf7a7580d707f60bd3baafd125fbee3a9d5