Bug 1770721: Clear Windows exception handler later in shutdown r=gsvelto
We want to clear this as late as possible to get the most accurate crash reporting.  In particular, we should clear it after ProcessRuntime is destroyed as its destructor does a lot of work and could lead to hangs that may still be processed by the Shutdown Hang Terminator.

Differential Revision: https://phabricator.services.mozilla.com/D147524

UltraBlame original commit: 85dc2fc8e905a7a72a0a3d232d019247ace4cfa5