Bug 1761849 - Removed 'ancestors' function and replaced the only use of it with a Pathlib equivalent r=firefox-build-system-reviewers,mhentges
While testing the change, I noticed that the old implementation also
considered the 'current' path as a part of all ancestors. As such, I
added an additional path component to the path that is immediately
discarded so that the behavior is consistent.

Differential Revision: https://phabricator.services.mozilla.com/D142273

UltraBlame original commit: 9fddfa1f41cf04ed5a5e1842c3bd27efe6f0bd05