Bug 1772447 - Add quick suggest results in chunks so the main thread isn't blocked too long. r=nanj
This modifies `UrlbarQuickSuggest.addResults()` so it adds entries to the
`_resultsByKeyword` map in chunks. Each chunk uses the idle queue of the main
thread.

I chose 1,000 as the chunk size. On my machine, a chunk with 1,000 entries
usually takes less than a millisecond, sometimes 1ms. A chunk with 10,000 takes
10x the time, ~10ms. Per the recent Slack thread, the latest total number of
keywords is 134,276, so that's 135 chunks of size 1,000. Seems OK...? I'm open
to suggestions.

Differential Revision: https://phabricator.services.mozilla.com/D148221

UltraBlame original commit: 847ac2732d90ec9ba561914a974349ae476abc97