Bug 1772088 - [devtools] Align DevToolsWorkerParent and DevToolsFrameParent around unregisterWatcher method. r=jdescottes
The `_connections` map is still not quite the same between the two,
and workers auto-destroy the JSWindowActor when the last connection closes.

But this auto-destroy looks meaningless? If the map is empty _destroy won't do much...

Regarding merging or sharing code between the two actors,
I don't have a clear picture on how to do so as there is subtle differences.
It might be clearer once content process targets are using a JSProcessActor.
It would introduce a third Actor, doing probably the same things again, in a yet different way!

Differential Revision: https://phabricator.services.mozilla.com/D148477

UltraBlame original commit: e594e3c885a2482a11ad8227ae88e10197670313