Bug 1770403, part 1 - Change initialization of mStorageAccessPermissionGranted on outer window, r=anti-tracking-reviewers,timhuang
mStorageAccessPermissionGranted was initialized to ShouldAllowAccessFor(...).
This had logic that has nothing to do with this the permission being granted.
A few bugs with my helpers were covered up by this default value, so I fixed the helpers in the same patch here.

This also required adding a new helper to the RejectForeignAllowList and exposing it to the Document.cpp file.
I'm not enthused about making that available, but we don't have a good component boundary for now, so I don't hate it.
Ideally the implementation of the rSA and hSA funcitons (and the like) would be in our component, but that is outside of this change's scope.

Differential Revision: https://phabricator.services.mozilla.com/D147867

UltraBlame original commit: 9e5ed810427c43d52b4c2e61ac4de66320c1744e