Bug 1769791 - Remove support for setting permissions in mozprofile. r=jmaher
The only permission that is set is allowXULXBL, and we don't want
to set that by default, because we don't normally allow XUL in
web content. It can be enabled on an individual test by adding
allow_xul_xbl = true to a test manifest, or by using pushPermissions
within the test.

This also removes the noxul test server, because no XUL is now
the default behavior.

Differential Revision: https://phabricator.services.mozilla.com/D148136

UltraBlame original commit: 670004b987ac7e78a204d942c7dfa3c450ab8ac3