Bug 1773905 - Remove prepared_frame_id and chase id r=gfx-reviewers,lsalzman
As part of the frame building work being done, the PrimitiveInstance
struct will change significantly (most of the state will be moved
to the primitive template, instance will be closer to an index +
some visibility state).

`prepared_frame_id` is no longer relevant (it can't be incorrect
now due to the way prims are reset). The chase is is not used or
maintained, and would need to be quite different after the changes
mentioned above anyway.

Differential Revision: https://phabricator.services.mozilla.com/D149025

UltraBlame original commit: 0405deb9402600c6da4cdf03705144e8857ac82c