Bug 1774408 - Don't use widget scale in AppWindow. r=tnikkel
This does fix the Pre-XUL skeleton UI scaling for me, but the others are
harmless and I can't repro the bug as described.

I _guess_ this can happen if you have window borders on Windows, since we were
mishandling those? But I don't know how to get those on Win10+.

The other scalings are symmetrical so they ended up working out in the end. In
any case this is the right thing to do and I'm ~sure it will fix the bug in the
reporter's configuration.

Differential Revision: https://phabricator.services.mozilla.com/D149449

UltraBlame original commit: 32f6c59157f35415243d6563003ef235b8a98d92