Bug 1744524: part 3) On the C++ side: dispatch a custom event when `clipboard.readText()` is called and allow the JS side to propagate whether the user clicked or dismissed the "Paste" button. r=edgar
This implements dispatching a custom "MozClipboardReadTextPaste" event
to the JS side, allowing the latter to show a "Paste" button. The JS
side may then report back to the C++ side, whether the user clicked or
dismissed the "Paste" button.
Combining these features is implemented in a following part and
`clipboard.readText()` is gated behind a pref in another following part.

The implementation bundles subsequent requests of `clipboard.readText()`
belonging to the same transient user activation timestamp.
For the first request of `clipboard.readText()`, a
"MozClipboardReadTextPaste" event is dispatched. As long as no response
for the event was received, further `readText()` requests are queued.
When the response is received, the calls are either resolved or rejected.
New calls following those, within the same transient user activation
period, are resolved/rejected too.

Differential Revision: https://phabricator.services.mozilla.com/D135335

UltraBlame original commit: 15610acc0c7ad18eb880b4fb3ee7476e8358bdac