Bug 1773650: Move final yield to end of function r=jandem
We generate very similar code when returning as when we fall off the end of a generator/async function. This patch unifies the two cases. We generate less bytecode this way. Also, OOMs or debugger-induced throws during the return will not be caught by try-catch blocks inside the function (although the outer reject catch will still see them; this is tested in `async/debugger-reject-after-fulfill.js`).

It's a bit weird because AsyncResolve is infallible in the spec but not our implementation, but I think this approach is most consistent with the semantics in step 3 of (AsyncBlockStart)[https://tc39.es/ecma262/#sec-async-functions-abstract-operations-async-function-start] (called by EvaluateAsyncFunctionBody). The entire body is evaluated before we check `result.[[Type]]` to decide whether to resolve or reject the promise.

Depends on D149469

Differential Revision: https://phabricator.services.mozilla.com/D149470

UltraBlame original commit: b7c7880c641033733260e65c766da83af94e7731