Bug 1774037 - Detect whether to use light or dark text masks on macOS. r=aosmond
macOS may apply differing amounts of font dilation depending on whether or not it
determines the text color is light or dark. We may end up applying the wrong amount
of dilation to text if we use a single mask texture for all colors. To work around
this, as in Skia and WebRender, we have to determine whether the text is light or
dark and then allocate separate mask textures for each case.

Differential Revision: https://phabricator.services.mozilla.com/D149494

UltraBlame original commit: 5097dc5f0e4ec87b89991f49a83bc7534fd4d1c7