Bug 1772347 - [devtools] Ensure that content process targets destruction is notified to the client. r=jdescottes
That, when the target is destroyed for any reason other than the process destruction or
when we stop watching for content process targets.
In these two cases, we will notify about the target destruction.
AFAIK, this is actually the only two reasons when a content process target is destroyed.
But it would be safer to assume that the target might be destroyed for some other reason
and have the client to be notified.

Depends on D148529

Differential Revision: https://phabricator.services.mozilla.com/D148137

UltraBlame original commit: a6d7efb82917eaa335a9b2dede096193687990be