Bug 1776074 part 2: Remove redundant assignment of mPrintObject->mFrameType. r=emilio
This patch doesn't change behavior.

The assignment here is working with a freshly-constructed `nsPrintObject`,
which was initialized via `InitAsRootObject`. Its `mFrameType` will have
already been assigned to `eDoc` in the constructor.  The only place in our
codebase where we might hypothetically reassign `mFrameType` is in
`InitAsNestedObject`, but that's the other init method -- not the one we're
calling here.

So: `mFrameType` will already be `eDoc` when we reach this assignment, and
the assignment is unnecessary.

Differential Revision: https://phabricator.services.mozilla.com/D150076

UltraBlame original commit: 37971cb5cdf3aa106858e71a2978d3f38c189be6