Bug 1776530 - Assert that APZUpdater::RunOnUpdaterThread() is not called directly from the updater thread. r=tnikkel
With the exception of the egde case scenarios in which
IsConnectedToWebRender() returns false.

Also add a comment warning that if RunOnUpdaterThread() were to be called
directly from the updater thread when connected to WebRender, the
implementation would be incorrect.

Differential Revision: https://phabricator.services.mozilla.com/D150309

UltraBlame original commit: ec0bd8f37b9abfd2d0ac7c22a7ef5cbfb7323980