Bug 1775480 - Log number of total and active renderers when EGL surface creation fails. r=gfx-reviewers,aosmond
The previous patch in this bug added crash annotations for the number
of total and currently active renderers. However, we are actually
interested in these values when the EGL surface creation fails, as
opposed to when we crash later on after failing to recover. This patch
adds the values to gfxCriticalNote at the time of the error.

Differential Revision: https://phabricator.services.mozilla.com/D150365

UltraBlame original commit: 9b7d03d3d4453680438df0d1db6c03a25dcfd6fd