Bug 1776649 - Don't require webdriver or geckodriver crates to be audited, , bholley r=webdriver-reviewers,supply-chain-reviewers,bholley,whimboo
This requirement meant that we had to publish the crates on crates.io
before they reached autoland. This didn't work well for the
geckodriver release process, since the binary releases are built on
mozilla-central, and needing to pre-upload the crates made it
impossible to guarantee that the binary release matches the uploaded
version.

Instead, add a step to the release docs to update this information.

Differential Revision: https://phabricator.services.mozilla.com/D150361

UltraBlame original commit: 812966d71698ff991910a561acb67506575444c5