Bug 1758789 - part1 : use an extra header to include API which are only available on win8+. r=jolin
Currently, we build with WINVER=0x601 (Win7), which means newer declarations in
mfmediaengine.h will not be visible. Also, we don't yet have the Fall Creators
Update SDK available on build machines, so even with updated WINVER, some of
the interfaces we need would not be present.

To work around this, until the build environment is updated, we include copies
of the relevant classes/interfaces we need.

Differential Revision: https://phabricator.services.mozilla.com/D144630

UltraBlame original commit: d7485e0cfe18ed6853e0669cf7bc794745713452