Bug 1775797 - Make resizers use frame-relative coordinates. r=TYLin
We no longer use the same code to resize windows and so (bug 1694061),
so keeping screen-relative coordinates is not really needed.

Instead, keep the size of the resized element around and do the math in
frame-relative coordinates, which is better behavior in presence of CSS
transforms.

Differential Revision: https://phabricator.services.mozilla.com/D150058

UltraBlame original commit: 5401aa8df48127eb7cba8b386961562f9455e611