Bug 1777211 - [devtools] Fix alignement of sub folders when picking a project root. r=bomsy
This css rules comes from ancient time and couldn't find the reason why this was added.
This is actually the immediate children of the project root which are misaligned.
That get an extra padding that make them look like they are in a sub folder.
I do not see the value of adding an extra padding. We still have some margin even after removing it.

Differential Revision: https://phabricator.services.mozilla.com/D150545

UltraBlame original commit: 3423a6a1b26ae575f6ee1a35a39e3042c429aa2d