Bug 1777215 - [devtools] Avoid expanding the tree manually in browser_browser_toolbox_debugger.js r=bomsy
In the following patches, this test started failing because of some
race condition between the test trying to expand the tree
and selectSource also trying to expand the tree.
We ended up we the tree being collapsed.

We no longer have to manually expand the tree as selectSource will now do it.

Differential Revision: https://phabricator.services.mozilla.com/D150547

UltraBlame original commit: 559f72945cbdaf1e9bbb4671baa4391801fd9652