Bug 1702806 - Fix issues with intermittient failures in places tests and due to the way the dialogs interact with PlacesTransactions. r=mak
There's are a couple of issues that this fixes, they are both based around the issue that because
tests run fast they tend to not give time for batches to complete. This can lead to incomplete
batches between tests, which tends to mess up subsequent tests or cause hangs.

In normal usage this is not an issue, as the user is not able to go as fast as tests.

The first issue is that in PlacesUIUtils.showBookmarkDialog we may resolve the dialog deferred
promise before the transaction batch has had change to complete.

The second issue is that the star UI runs batches, and that again we do not wait for the batch
to be complete. Working an await into the star UI/tests is more difficult due to the way that
code is structured. With a planned switch to non-instant apply, it also feels too complicated
to be worth it at this time. As a workaround for this issue, we ensure that the transaction
queue is cleared between each test run via a registerCleanupFunction in the head.js file.

Differential Revision: https://phabricator.services.mozilla.com/D147535

UltraBlame original commit: 5ff5566688462dc0ae916616b784650050945581