Bug 1777222 part 5 - Add a small cache to optimize NewPlainObjectWithProperties. r=jonco
With TI we had a cache for this but we got rid of it when we removed TI. This patch
adds back a simpler version of it that caches the four most recently created
shapes. This cache has a good hit rate on the web and on Kraken's json-parse-financial
test.

Differential Revision: https://phabricator.services.mozilla.com/D150558

UltraBlame original commit: 585e41034e1e8d3f403ddf474e828e153b039b04