Bug 1771858 - Handle interrupt in AsyncModuleExecutionFulfilled; r=jonco
In AsyncModuleExecutionFulfilled, it's possible that the call to GatherAsyncParentCompletions
will fail if we are interrupted. The current code assumes that this is only reachable as a
result of an OOM. This changes the assertion to an if statement to handle being interrupted.

Differential Revision: https://phabricator.services.mozilla.com/D147954

UltraBlame original commit: b30d65d0e2eeb96d84669e03c88cb9ab9d7adcfd