Bug 1775327 - Part 3: Do normalization for NormalizedTiming(). r=firefox-animation-reviewers,birtles
This implements the normalization of the specified time, defined in
[web-animations-2]:
https://drafts.csswg.org/web-animations-2/#normalize-specified-timing.
However, it is possible to update this, based on the spec issue:
https://github.com/w3c/csswg-drafts/issues/4862.

For now, we just do normalization for delay, end delay, and
iteration duration based on the end time. And make sure the end time is
equal to the timeline duration.

Differential Revision: https://phabricator.services.mozilla.com/D149685

UltraBlame original commit: 71bbfaecf629a3d73c72600d35c0ac2965516d60