Bug 1742438 - Prep Part 2: Create an array of references for ScriptLoadInfos; r=asuth
This patch introduces a temporary array for the purposes of this refactoring. First, I'll
demonstrate the new loading strategy using the nsTArray datastructure, before moving everything over
to a ScriptLoadRequestList and ScriptLoadRequest datastructure. The impact of this change will only
exist for the course of this bug.

Differential Revision: https://phabricator.services.mozilla.com/D145442

UltraBlame original commit: ac92b01963abac829fece861ee7011e26b78571f