Bug 1742438 - Prep Part 10: Make TriggeringPrincipal optional for ScriptFetchOptions; r=jonco
I didn't realize that this would only be used in the DOM script loader, so this is now optional. In
a later stage we might move this over to the ScriptLoadContext as it looks like this might need to
propogate through the module script tree.

Differential Revision: https://phabricator.services.mozilla.com/D146172

UltraBlame original commit: 66940c72c03b0c284d08b5981be26816641a0bed