Bug 1543337 - Split touch pointer event into two, r=webdriver-reviewers,whimboo,jdescottes
For touch type pointers gecko doesn't support all the properties in pointer-events (and
it's not even very clear to me that they all make sense; how can one talk about the twist of
a finger, for example?). So split out a version of the tests that only test the properties
which are supported, so that we can regression-test the basics without failing due to missing
support at the engine layer for more complex properties.

Differential Revision: https://phabricator.services.mozilla.com/D139121

UltraBlame original commit: ac6f0d293d47ee563ec91fd636989bd6db4d3890