Bug 1770498: Use the final channel URI for SubDocument types also r=ckerschb
Testing indicated that for a SubDocument, the loading principal
is that of the parent document, but the final channel uri is
the uri of the subdocument.  In this situation, use the final
channel uri just as we do for a Document.

Differential Revision: https://phabricator.services.mozilla.com/D150586

UltraBlame original commit: 8f6e60b4c2212e341af17e8952e9c7b8cbd55767