Bug 1770498: Update the browser_canvas_rfp_exclusion.js test r=timhuang
Previously, this test would open a page and then enable/disable
RFP while the tab was open, writing and reading the canvas as
it went.

Now that CookieJarSettings gets initialized when a document loads
and stays constant for the lifetime of the document, we cannot enable
and disable the prefs and expect them to continue to behave correctly.

We do a 'pre-test' to populate a canvas and then get the true image
data for it. Then we set the preferences prior to loading future
documents and compare the extracted canvas data to the known-correct
data to see if we are correctly spoofing it or not.

Differential Revision: https://phabricator.services.mozilla.com/D150592

UltraBlame original commit: 20d198e82c502c340ad4a8fa3ba3fce0fbf7c123