Bug 1779043 - [devtools] Fix jumping to media query in inline sheets. r=jdescottes.
We directly pass the editor stylesheet in the `location` parameter that is passed
to `jumpToLocation`.

`jumpToLocation` wasn't covered in any test, so we modify `browser_styleeditor_media_sidebar.js`
to add coverage for it.
This is also testing Bug 1779180 which is in the same stack.

Differential Revision: https://phabricator.services.mozilla.com/D151597

UltraBlame original commit: 95e34a0fbb39c2bcf94e60c06f524ec99743f4fd