Bug 1773324 - Remove template from ErrorAllocator, add ErrorContext::getAllocator(), rename OffThreadErrorContext::linkWithJSContext() r=arai
ErrorAllocator was only ever used with ErrorContext, so no need for templating.

linkWithJSContext() better describes what it is doing, since JSContext needs its
errors_ field filled in, at least until we can replace it with a different
allocator for functions that are called during compilation.

getAllocator() is a convenience so we don't have to add a separate allocator
parameter everywhere we pass an ErrorContext and also need an allocator.

Differential Revision: https://phabricator.services.mozilla.com/D151609

UltraBlame original commit: e31782ad0907fbcbbe1c826dddf04d862df30979