Bug 1779173 - [devtools] Fix call to isMinified in Debugger to prevent performance issue when opening big single-line file. r=jdescottes.
Changes in Bug 1754407 caused a performance regression as it now triggered
syntax highlighting of big files, which we avoided before.
This is because we don't call `isMinified` with the expected type of data, causing
the file to not be seen as minified, and thus highlighting it.

Differential Revision: https://phabricator.services.mozilla.com/D153140

UltraBlame original commit: d28b00fe705ba7c6855831834e55e515031e4b2a