Bug 1776853 - [devtools] Add BrowserToolbox-specific toolbar. r=ochameau,jdescottes,fluent-reviewers,flod.
This adds an additional toolbar in the Browser Toolbox which will contain specific
tools and options.
At the moment we only display a couple input button to be able to switch from
Parent process only to multiprocess mode.
We remove the similar UI in the iframe picker and adapt the existing test.

A trait is added to not show the toolbar when debugging server where we wouldn't
get the `isSwitchingMode` property in `onTargetDestroyed`, as this can cause
misbehavior in various tool when switching between different modes.

Differential Revision: https://phabricator.services.mozilla.com/D150575

UltraBlame original commit: c97931892c2363a033e9278d8eb777d8c62ee392