Bug 1782488 - [devtools] Clear console service cache in browser_console_modes.js for non-fission case. r=jdescottes.
In the test, when fission isn't supported, we use Cu.reportError within
a SpecialPowers task, which puts the message in the console service cache.
If the test runs multiple time, the next iteration will run the test
with Fission support, which will consume cached messages, making the reportError
message appears more than we want.
This isn't a problem we can face in normal usage as when Fission is not supported,
we don't retrieve cached messages.
In order to fix the test failure, we simply clear the console service cache from
the content process.

Depends on D153327

Differential Revision: https://phabricator.services.mozilla.com/D153353

UltraBlame original commit: 32b5787d9b6761f6d55ead837c3777a1a82ea06b