Bug 1782717 - Fix back/forward menu tooltip string for current page item to indicate it will reload page. r=Gijs,flod
The back/forward menu shows tooltips when hovering menuitems. For the
current page, the tooltip says Stay on this page, but the effect of the
menuitem is actually to reload the current page. So change the tooltip
string to reflect that the page will reload when the item is clicked.

Differential Revision: https://phabricator.services.mozilla.com/D153468

UltraBlame original commit: 34e61e1476efb312c8825278cbdd768973574f87