Bug 1761291: create wrapper to launch Firefox in Private Browsing mode r=rkraesig,glandium
This allows us to fix a bug where when our current Private Browsing shortcuts are pinned to the Start Menu, they use the regular Firefox Visual Elements (which is the non-Private Browsing logo). I tried to make this as minimal and braindead as possible.

Differential Revision: https://phabricator.services.mozilla.com/D151538

UltraBlame original commit: 84c998372e2fe77b250682628998917fd060dbf5