Bug 1772205 - Implement a special version of Http2Stream that will be used for tunneling. r=necko-reviewers,kershaw
The stream will behave as HTTP/2 stream for the HTTP/2 session to the proxy  as well as a SocketTransport for the nsHttpConnection through the tunnel. nsHttpConnection will be responsible for sending the CONNECT request the same way as it does in case of a HTTP/1.1. proxy.

Differential Revision: https://phabricator.services.mozilla.com/D149978

UltraBlame original commit: ea5625892397dd502be2f6d3045c8f9eacbc3f6e