Bug 1782852 - part 2: Disable Gecko specific commands which are disabled only in the Nightly channel in early beta too r=smaug
This is a trial to get reports of breakage of web apps in the wild from (early)
beta users. Currently we don't have any bug reports from Nightly users of
disabling `contentReadOnly` and `insertBROnReturn` commands, but according
to the telemetry in the Beta channel, there are some web apps using them
(perhaps, for feature detection or one of finger prints) because we get over
1k pings per day.

`contentReadOnly`:
https://telemetry.mozilla.org/new-pipeline/evo.html#!aggregates=True%2520percentage&cumulative=0&end_date=2022-07-25&include_spill=0&keys=!__none__!__none__&max_channel_version=beta%252F103&measure=USE_COUNTER2_DOCUMENTEXECCOMMANDCONTENTREADONLY_DOCUMENT&min_channel_version=beta%252F100&processType=*&product=Firefox&sanitize=0&sort_keys=submissions&start_date=2022-06-27&trim=0&use_submission_date=0

`insertBROnReturn`:
https://telemetry.mozilla.org/new-pipeline/evo.html#!aggregates=True%2520percentage&cumulative=0&end_date=2022-07-25&include_spill=0&keys=!__none__!__none__&max_channel_version=beta%252F103&measure=USE_COUNTER2_DOCUMENTQUERYCOMMANDSUPPORTEDORENABLEDINSERTBRONRETURN_DOCUMENT&min_channel_version=beta%252F100&processType=*&product=Firefox&sanitize=0&sort_keys=submissions&start_date=2022-06-27&trim=0&use_submission_date=0

If we won't have any regression reports with this, for example, it's used only
for a finger print, perhaps, we could try to roll out to disable these commands.

Differential Revision: https://phabricator.services.mozilla.com/D153721

UltraBlame original commit: 99538761dfaf1e3e27bed6b622fd37f292aeb3bf