Bug 1783160 - [devtools] Use private fields and methods in BrowserToolboxLauncher. r=jdescottes.
Tests were using the internal `_dbgProcess` and `_dbgProfilePath` properties, which
we migrate to private properties.
We now pass those properties as part of the `run` event that those tests are
setting up, so we can run assertions on those object.
Finally, we take this as an opportunity to clean `browser_dbg-chrome-create.js`.
Since it was the only callsite using `onClose`, and given that we directly call
the `close` method, we can remove the `onClose` from the signature of `BrowserToolboxLauncher`.

Differential Revision: https://phabricator.services.mozilla.com/D153733

UltraBlame original commit: d0829e7bef084d0b294883806ea36e078a57ec14