Bug 1783608 - [devtools] Make `evaluateInBrowserToolbox` throw when passed expression does throw. r=jdescottes,perftest-reviewers,sparky.
Since the test is using `evaluateJSAsync` with `mapped` set to `{await: true}`,
if the Promise returned by the expression rejects, the response will have a
`topLevelAwaiRejected` property set to `true`.
In such case, we directly throw from `evaluateInBrowserToolbox`, so the test
will fail.
We also cover the case where the passed function isn't async, in which case the
response will hold a `exceptionMessage` property that we can use.
Finally, we take this as an opportunity to remove duplicated function, as well
as removing a try/catch block in one of the expression since it was also hiding failures.

Differential Revision: https://phabricator.services.mozilla.com/D153919

UltraBlame original commit: b347201c43084e02792595b337bc4074e19bdc07