Bug 1783717 - Use hash chaining for DrawTargetWebgl cache entries. r=aosmond,bradwerth
This implements a rudimentary form of hash chaining for cache entries that reduces
the number of entry comparisons by about an order of magnitude. It's a bit of a bother
to use an actual hash table here since right now the code expects to be able to have
random access to cache entries in the list, and the MFBT hash table is not quite set
up for this. Instead, this just adds a reasonable fixed number of buckets to bring
the list size better under control.

Differential Revision: https://phabricator.services.mozilla.com/D154101

UltraBlame original commit: 934c29e602e6ab6bd884db753dc793dee10eec10