Bug 1780792 - Move the remaining buffer logic in Device.cpp into Buffer.cpp. r=jimb
Having the code in the same place makes it easier to follow. This made me realize that the validation of aMode in mapAsync has to move to the device side (fix coming in a followup).

Depends on D151631

Differential Revision: https://phabricator.services.mozilla.com/D151632

UltraBlame original commit: b4210142bf82d4c0c32c374d8cc2937bafe2ad2a