Bug 1782578 - Put the img shadow DOM behind a preference r=nordzilla
This code will most likely be removed, but for now I am putting it
behind a pref so that it won't be activated when the modal is being
created.

See Bug 1782582 for the removal of the code.

Differential Revision: https://phabricator.services.mozilla.com/D153393

UltraBlame original commit: a77a405173fff17b829da8f1200c3099d5b3e99b