Bug 1784251. Allow widget/tests/browser/browser_test_swipe_gesture.js to test swipe to nav on linux. r=hiro
We implement SendNativeTouchpadPan/SynthesizeNativeTouchpadPan (like Windows does for this test).

I tried to use the existing functions SendNativeMouseScrollEvent/SynthesizeNativeMouseScrollEvent which are implemented on Linux and which are what we use for mac on this test, but it's already used for other stuff and it would be very clunky to overload it to make it work for this too.

I didn't see any way to "tag" the gdk events with more info, so making the observer notifier work was clunky. Similarly for getting the phase start/update/end work.

Differential Revision: https://phabricator.services.mozilla.com/D154386

UltraBlame original commit: d09cd2d80dbce94b4491fdc832a7ea8a285c4d03