Bug 1784287 - Use input margin for screen size expansion. r=tnikkel
This is a follow-up to bug 1783500. The existing expansion for the
screen area works great on Windows and so on, but on macOS it can
conceptually cause a menulist to go off-screen, because of this margin
used to move menulists to the left:

  https://searchfox.org/mozilla-central/rev/f655bdf6b4bf01b42609750ab94fc37635397260/toolkit/themes/osx/global/popup.css#85

Instead we should do the same as that bug did, and use the
input-region-margin, which is the amount of space that has no content
(that is, that contains the shadow and so on) and is zero on macOS
(because shadows on macOS are drawn by the OS unlike on Windows /
Linux).

This required extra test changes so it was worth getting it reviewed
separately.

Differential Revision: https://phabricator.services.mozilla.com/D154401

UltraBlame original commit: 8e225be30ad27b166465600ed04635b15fd30e52