Bug 1784940 - Acquire/release locks in gfxFont::Measure fewer times. r=jfkthame
gfxFont::Measure would acquire/release a read-only lock on
gfxGlyphExtents::mLock for every glyph in the text run. This patch makes
it so that we once the lock once. We may need to populate the
gfxGlyphExtents hashtable once, where we need to upgrade to a read-write
lock.

Differential Revision: https://phabricator.services.mozilla.com/D154816

UltraBlame original commit: 866e80f446586dea16695d94dc8e672fd1b652d5