Bug 1779005 - Use mLoadFailureCount as an indicator for PAC laod, r=necko-reviewers,valentin
Since IsLoading() is false when PAC request fails, we have to use something else to indicate that PAC is not available.
mLoadFailureCount can be used in this case. It only reset when PAC request is succeeded and when reloading a PAC.

Differential Revision: https://phabricator.services.mozilla.com/D154907

UltraBlame original commit: a9cbfd43cf72131637a9dbc77097a2346a7a1342