Bug 1783601 use GetChannelResultPrincipalIfNotSandboxed() for media resource principal r=chunmin
When the media resource is loaded as a document, the response from the initial
document load gets reused, as an optimization, as an emulated load for the
resource of the media host element in the generated HTML document.
https://searchfox.org/mozilla-central/rev/5644fae86d5122519a0e34ee03117c88c6ed9b47/dom/html/VideoDocument.cpp#114

Sandboxing from an iframe in a parent document can cause the document load
to have a null result principal for the sandboxed HTML document
https://searchfox.org/mozilla-central/rev/778216db0ad397de415743ec9845e3e3dc6e7112/caps/nsScriptSecurityManager.cpp#316,332-333
but this principal differs from that for subsequent loads of the same media
resource from the host element.
https://html.spec.whatwg.org/multipage/browsing-the-web.html#read-media
The change in principal was incorrectly interpreted as a prohibited
cross-origin redirect.

Ignoring sandboxing when obtaining a principal from the document load provides
the same principal as would be obtained by the load from the host element.

GetChannelURIPrincipal() is not suitable because "data:" URLs should inherit
their principal, as tested by
testing/web-platform/tests/html/canvas/element/manual/imagebitmap/createImageBitmap-serializable.html

Differential Revision: https://phabricator.services.mozilla.com/D154043

UltraBlame original commit: f64890d4733bc95ba76e6eff43df0d421723b252