Bug 1746699 - Avoid UB when allocating JSStrings and BigInts r=tcampbell
Add emplace methods to JSString and BigInt types to reduce amount of C++
undefined-behaviour by running the constructor of the concrete derived type.
This is in contrast to previous behaviour of allocating a gc::Cell and then
simply casting to the final type. We still first allocate as a Cell from the GC
and then use placement-new to initialize as the correct type.

Differential Revision: https://phabricator.services.mozilla.com/D134095

UltraBlame original commit: ac7a2cfb2cd0d34fccc0dcc04225d9ef3bbdd20e