Bug 1694993 - Part 4: Copy session storage before loading, and only for newly created windows, r=smaug
Previously we would copy session storage even if we were not opening a new
window, meaning that a targeted load could re-trigger a copy. This was not
specified in the standard so is being changed to only copy when a new window is
created. In addition, the copy was moved before navigaton starts, again for
more consistency with ordering for the standard, such that things like
`javascript:` URI loads will oberve the up-to-date session storage.

Differential Revision: https://phabricator.services.mozilla.com/D154566

UltraBlame original commit: 54e339586d2682bf180d133f6c933e32a0c36927