Bug 1786830 - [devtools] Display container queries on top of rules. r=jdescottes.
This is a first basic implementation that we might extend in the future
(for example by adding a way to select the container node when clicking
on the container name).

A test is added to ensure we display container queries as expected, even
when dealing with unknown containers, or with nested layer/media/container rules.

Depends on D155602

Differential Revision: https://phabricator.services.mozilla.com/D155469

UltraBlame original commit: 4c76664026b55d57999e109b5bc5429d986df9ab