Bug 1786465 - Convert OriginPrivateFileSystemChild to a standard actor; r=dom-storage-reviewers,jesup
The send methods which need to be mocked in tests are declared in
OriginPrivateFileSystemChild the same way as POriginPrivateFileSystemChild send
methods, just without the Msg suffix and virtual. The relevant
POriginPrivateFileSystemChild send methods are hidden in
OriginPrivateFileSystemChild, so they can only be used internally for the
forwarding.

Differential Revision: https://phabricator.services.mozilla.com/D155957

UltraBlame original commit: 16b9fd20cbf4e3f0b106940f045819d829c97826