Bug 1785063 - Obey contain-intrinsic-size in select multiple. r=emilio
When there is size containment, if the contain-intrinsic-size is not none,
we need to use the available contain-intrinsic-size to determine the
intrinsic inline and block sizes.

This patch also updates and adds a couple of sub-tests in contain-intrinsic-size-009.html.

Differential Revision: https://phabricator.services.mozilla.com/D155051

UltraBlame original commit: d9636db75bddb4bbf8f8e81c1947612a3aed83d2