Bug 1787461 - Change Feature Callout's overlap of parent elements to 17px. r=mviar
Feature Callouts are absolute positioned, but their coordinates are calculated
from those of the parent element given by a selector property on the message.
They were originally adjusted to overlap by 15px, but that was changed to
0.15x - y, where x is the parent width/height and y is the callout's arrow size
(12px). But using a scaling factor may cause unexpected results where parent
elements are unusually large, so this patch switches the overlap amount to 17px.
In the future, we could consider adding a property to the schema so that each
feature tour screen can specify an overlap amount to override the default 17px.

Differential Revision: https://phabricator.services.mozilla.com/D155750

UltraBlame original commit: b6d4cc392199eac68b7641fe6ce2be73a84381f6