Bug 1788858 - Remove HOST_MAJOR_VERSION. r=firefox-build-system-reviewers,necko-reviewers,andi
`HAS_CONNECTX` is not used in `netwerk/protocol/http` as of bug 1689604.
In NSPR, it is used to enable the use of `connectx`, which is available
since macOS 10.11, and we target 10.12, which means it's always
available, so we don't need to check for it based on the target version
(and checking the host version was wrong, too).

Differential Revision: https://phabricator.services.mozilla.com/D156275

UltraBlame original commit: b1b9abc406dffa627b1bf3a5383db0473b581281