Bug 1789349: Disable test test_bug706406.html. r=emilio
This is testing the legacy <maction> element, a feature currently
disabled by default and intended to be removed in the future. The test
tries to enable it, but that does not always work. So let's disable the
test for now, it will be removed in the future.

Differential Revision: https://phabricator.services.mozilla.com/D156528

UltraBlame original commit: 529bdf2c8afd9f21e5bfffd25ff6c88b63326680