Bug 1788115 - Do not enforce 63 character limit for IDNA labels r=necko-reviewers,dragana
According to https://unicode.org/reports/tr46/#ToASCII we have the option of
setting VerifyDnsLength = false for the algorithm. This seems in accordance
to the URL spec and the rest the browser implementations.

The label size is arbitrarily bumped up to 256.

Differential Revision: https://phabricator.services.mozilla.com/D156398

UltraBlame original commit: 6a46724141de9094c0f59ce4c812b028aa915bc5