Bug 1787115 - Allow Colorways onboarding modal to restore user-installed themes. r=pdahiya
This is the simplest approach I could think of to allow the welcome screen to
revert back to any theme the user had installed initially. The reason this
wasn't working previously is just that the surface could only retrieve built-in
addon IDs. If we reserve a property for an active non-built-in theme, then the
behavior will be the same for built-in themes and user-installed themes.

Differential Revision: https://phabricator.services.mozilla.com/D155769

UltraBlame original commit: c4b72690ced0cb82fdd1fce274f15fe6fced9f96