Bug 1786185 - Migrate XMLPrettyPrint from DTD to FTL. r=gregtatum,smaug,fluent-reviewers,flod
The synchronous DocumentL10n instance that's created here is not exposed on the root document, as that is the user-defined XML document. The localization root that's connected to it is in a closed shadow root.

This localization won't update on locale changes, but that matches what this view was previously doing.

Differential Revision: https://phabricator.services.mozilla.com/D156544

UltraBlame original commit: 5c9f3a2ee4ad666e4a5b0bce68a1023cc91363eb