Bug 1692653 - Fall back to locale code without region when resolving language names in Accept-Languages, r=Gijs,preferences-reviewers
If a locale code with region (e.g. it-IT) is not defined in language.properties, we display an empty language name.
In this dialog, we can still show a language name if it matches a known locale code without the region (e.g. it).

Differential Revision: https://phabricator.services.mozilla.com/D157113

UltraBlame original commit: 08963f8a53c79c23ae8c6001a03f236b560cd835