Bug 1787520 - Avoid invalidating render targets on Mali-G78 devices. r=gfx-reviewers,jrmuizel
On the Pixel 6 family devices we are seeing frequent image corruption
issues on some websites, which started following the Android 13
upgrade. This can be avoided by not invalidating the no-longer-needed
render targets at the end of each render pass. This is only an
optimization anyway, so is safe to skip.

The Android 13 update upgraded the Mali driver from version v1.r32p1
to v1.r36p0. As we did not encounter this bug prior to the Android 13
update, this patch only applies the workaround on driver versions
v1.r36p0 and above. It's possible other GPUs than the G78 are also
affected, but for now we limit the workaround to just this GPU. We can
re-evaluate if and when we receive bug reports from users on other
GPUs.

Differential Revision: https://phabricator.services.mozilla.com/D157116

UltraBlame original commit: ce1c3f8552f73ac1adae861acfac27c727a62ba2