Bug 1782085, part 3 - Add tests for partial FedCM API, r=timhuang
There is a lot here, and yet it isn't the best test coverage. My testing strategy is to use these as sanity/regression checks until we have enough implemented to use sync'd WPTs.

The tests are all plain mochitests that use example.com as the RP and example.net as the IDP. They set a cookie in the IDP, tell the manifest endpoint what test is coming, then requesting a credential. The manifest endpoint can use the test name provided to point the browser to a particular set of test endpoints (server_X_accounts.sjs and server_X_idtoken.sjs).

This also includes a new preference that is needed to test the API. It bypasses the 120s timeout for all returned promise rejects in the spec. Otherwise testing this would take forever.

Differential Revision: https://phabricator.services.mozilla.com/D155720

UltraBlame original commit: a154b86a79f3840ff46ebef475774060303d9e57