Bug 1789457 - Move NativeObject check after cache lookup r=jandem
Initially we explored caching this check on the shape, which still seems like
a decent idea. However, we can also just sidestep the problem for the most part
be moving the check after the cache lookup, since we can't get a cache hit if
it's not a NativeObject anyway, and we're hoping that a cache hit should be the
dominant path in most cases.

Differential Revision: https://phabricator.services.mozilla.com/D156614

UltraBlame original commit: 080d321a6baf6a4ee27e3d08110ebefc3b9cb4e3