Bug 1790487 - Change Colorways tooltip and aria-label r=aminomancer,fluent-reviewers
I'll wait to land this until next week, so that we're not putting in new strings right at the end of Nightly.

I've tested this with MacOS VoiceOver & NVDA on windows and the behaviour is pretty good. VoiceOver will repeat the colorway group & name a few times which is annoying, but it's an existing problem we haven't found a tidy solution for yet.

Differential Revision: https://phabricator.services.mozilla.com/D157369

UltraBlame original commit: 1d5f095c02d2366a8e591c26fee1d305f1f078a4