Bug 1791518 - Don't rely on the raw output from clang-format in mozlint. r=linter-reviewers,andi
Because of a bug in clang-format, its output is not always 3 lines per
warning. So instead of relying on its raw output, use the xml
replacement output, and produce our own output, in the form of a diff.

Differential Revision: https://phabricator.services.mozilla.com/D157717

UltraBlame original commit: 0ed2c0142c66e4dfa2fab097581721fcb1eb7cf4