Bug 1778450 - Fix tab navigation in toolbar when Firefox View button is not present. r=Gijs
With the addition of the Firefox View button/tabstrip, there is a new
tabstop that is present at the start of the TabsToolbar. When the
Firefox View button is not present, this new tabstop causes tab
navigation to skip over the selected tab and instead focus the new-tab
button. Then, when trying to navigate backwards from the selected tab,
this tabstop at the front of the TabsToolbar forces us to re-focus the
new-tab button. This is because the tabbrowser-arrowscrollbox-periphery
container has zero width/height which causes the tabstop inside of this
container to be skipped over, which then focuses the new-tab button
instead of the selected tab.

This patch allows zero width nodes to be skipped instead of rejected.
This allows the tabstop within the zero width periphery container to be
the first valid element when navigating forward from the tabstrip,
instead of the new-tab-button that is outside of the <tabs> element.
This restores the previous behavior of a selected tab being the first
focusable item when navigating forward from the end of web content, if
there are no items in the tabstrip.

Differential Revision: https://phabricator.services.mozilla.com/D157351

UltraBlame original commit: e231687609e3639101b442ff14380e858480c004