Bug 1792063 - Fix menuhover background computation in some gtk themes. r=stransky
When figuring out menuhover background, make sure not to come up with
the exact same color as the menu background. This ensures we also try
gradients / background-images before giving up.

This is only an issue with the dark theme in this case, because
otherwise we use native drawing.

Differential Revision: https://phabricator.services.mozilla.com/D157963

UltraBlame original commit: 0d1e3f18731745e0e573bd7519ac70359e04fca9