Bug 1793609 - [Linux] Wait for OnSizeAllocate while the window becomes visible before updating window size. r=stransky
The first size-allocate is likely (or guaranteed) to be of the size at
the time of the first NativeShow(true) call. Any resize between show and
the first size-allocate will be accounted for in a second size-allocate
signal. If we update and dispatch the window size at the beginning of
such a resize, we end up reverting back to previous size from the first
size-allocate signal before dispatching the size for the resize again
from the second size-allocate signal.

Differential Revision: https://phabricator.services.mozilla.com/D158580

UltraBlame original commit: 75c1403f58f79d1abd43d33fdd1beb36db9367c6