Bug 1794654 - Wait for cache update before comparing bounds to content. r=morgan
We first retrieve accessible bounds and then pass them to a content task to compare against content layout bounds. We are not guaranteed to have the cache in sync in that manner, most tests will wait for a content rAF but that is not sufficient for the parent cache's state.

Instead we need to pull the content layout bounds (since those are guaranteed to be in sync via rAF) into parent and then wait for the a11y cache to catch up.

Differential Revision: https://phabricator.services.mozilla.com/D159080

UltraBlame original commit: c6e6251e298892f132124cd8197cb4df13878195