Bug 1791881 - Don't crash in buffer_destroy if the buffer is in an invalid state. r=jgilbert
Once https://github.com/gfx-rs/wgpu/pull/3094 is merged, unallocated and freed handles will panic in wgpu-core so we don't have to do it here. In the mean time it will produce the wrong error but still fail safely. DestroyError::Invalid means the handle exists but is not is in an invalid state, for example if the buffer was created with invalid parameter like in this bug's test case.

Differential Revision: https://phabricator.services.mozilla.com/D159054

UltraBlame original commit: a6870c0b803a4073ecdfb3c57ec849fcf60ce581