Bug 1794841 - wasm: Avoid unnecessary allocations in TagType. r=jseward
An exception handling tag `(tag $name)` will nearly always have at
least one param for communicating the pointer to the thrown exception.

We should increase the inline storage in this vector so that we
don't have to allocate in this case. We do this already with the
ValTypeVector used in this class.

Differential Revision: https://phabricator.services.mozilla.com/D159191

UltraBlame original commit: 747d8347abf1cc02324efaff671c76dc7c0f2696