Bug 1793523 - Replace IPC's use of waitpid with waitid to avoid racing with the crash reporter. r=nika
When IPC is waiting for a process to exit at the same time the crash
reporter is trying to attach to the main thread with `ptrace`, both of
them will use `waitpid` for their respective state changes, and these
calls can race: the ptrace-stop is delivered to IPC, while the crash
reporter and crashed child process hang forever.

This patch changes IPC to use `waitid` with `WNOWAIT`, which allows us to
check the process status without side-effecting it; if it is an exit, the
call is reissued without that flag to obtain the status as before.

(This may not be a problem on macOS, where we use Mach APIs instead of
`ptrace`, but `waitid` has been in POSIX long enough that we should be
safe using it on all Unix platforms.)

This patch also adds some logging when child processes exit with a
failure status or due to a signal.

Differential Revision: https://phabricator.services.mozilla.com/D159186

UltraBlame original commit: 0f1715b3e25d08079cfd214727615ba999933da5