Bug 1795780 - Add the ability to disable a browser migrator at runtime via a pref. r=NeilDeakin,Gijs
It seems that at build time, we already make decisions about what migrators we register
with XPCOM in browser/components/migration/components.conf. This doesn't, however, give
us the ability to turn off a migrator that happens to be misbehaving, nor make it easy
for us to do development of a new migrator without exposing it to the migration dialog
(without doing binary recompiles).

This adds another layer to the enable-ment of our migrators so that they must both
be registered AND enabled in order to appear in the import dialog.

Differential Revision: https://phabricator.services.mozilla.com/D159544

UltraBlame original commit: 500284b778624fc57597f6a9ca46ee66c989480d