Bug 1794757 - autoplay dialog should wait for translation to finish before showing and determining size to avoid buttons being cut off, r=kpatenio
The sitePermissions code is reused, and all dialogs work OK except the autoplay one.
This is because the autoplay code adds extra content not present in the other
dialogs, and localizes it.
The SubDialog code relies on determining the size of the dialog once the
mozSubDialogReady promise resolves, which it does before fluent puts in l10n data
for the content that was just added. This patch fixes that.

It also, as a driveby, correctly pauses and resumes observing l10n attributes
in the doc when making modifications, to avoid fluent doing a second pass for
the same content.

Differential Revision: https://phabricator.services.mozilla.com/D159543

UltraBlame original commit: 878883e92b09027c069df023c32ae6fae815f72e