Bug 1727820 - Change aboutDialog-appUpdater to accomodate changes to AppUpdater r=bhearsum
In a previous patch in this stack, AppUpdater was changed to remove the checkForUpdates function. This change updates aboutDialog-appUpdater to accomodate this change.

It was probably incorrect for this function to be called previously. Calling into checkForUpdates() is functionally similar to calling into check(), except that checks are bypassed that probably shouldn't be bypassed.

Differential Revision: https://phabricator.services.mozilla.com/D159299

UltraBlame original commit: 6d8cf5f78de88791e9d0b056caf059f8d1e659e6