Bug 1796590: Remove fuzzy-if(appleSilicon) annotation for css-grid/grid-item-video-stretch-001.html. (no review, just updating reftest manifest to reflect reality)
Looks like this test used to catastrophically fail on this platform, per this
annotation; and that failure seems to have been fixed by bug 1792257.

DONTBUILD when this lands on autoland, because it won't make a difference
since we don't seem to schedule tasks for this platform on that branch.

Differential Revision: https://phabricator.services.mozilla.com/D160390

UltraBlame original commit: 780d2d1d736f3eac0ef9745983822b498daa8e84