Bug 1797755 - Part 4: Remove option to set maximum mark stack capacity in release builds r=sfink
Currently we have a GC parameter that allows setting a maximum mark stack
capacity. This is only ever used by test code, not in the browser. This
requires extra unnecessary work in release builds if we move to a different
stack representation as we won't be able to fold the comparison into the
current capacity check as we do now.

The patch makes this feature condtional on JS_GC_ZEAL.

Depends on D160526

Differential Revision: https://phabricator.services.mozilla.com/D160527

UltraBlame original commit: f207beed6252d85e5a64cc768eaef9d8a47b4912