Bug 1745758 - Add DNR rule evaluation engine r=rpl
This patch provides a rule evaluation engine for the DNR extension API,
using a naive linear API to evaluate rules for a given request.

The main purpose of this is to offer DNR functionality for relatively
simple use cases, as the more complex ones are likely going to use the
blocking webRequest API instead.

Differential Revision: https://phabricator.services.mozilla.com/D156233

UltraBlame original commit: 588516a44335200658392ceaa49e6cd6a2632aec