Bug 1797750 - Part 10: Correct compartment/realm handling in WrappedFunctionCreate. r=mgaudet
Move the `AutoRealm` in `WrappedFunctionCreate` into a separate block to ensure
we're back in the original realm when we call `CopyNameAndLength`. Without this
change an error thrown `CopyNameAndLength` can escape the wrong realm to resp.
from the ShadowRealm.

Also add some `cx->check()` calls to ensure arguments are in the correct compartment
in `GetWrappedValue` and `WrappedFunctionCreate`. And add `cx->compartment()->wrap()`
in some places which we're using the wrong compartment.

Differential Revision: https://phabricator.services.mozilla.com/D160513

UltraBlame original commit: a581812a6162a959d263bfa76bbdd4e92077add8