Bug 1797155 - Part 1: Don't walk the #list children directly in OverflowableToolbar.#moveItemsBackToTheirOrigin. r=sclements
This patch makes it so that OverflowableToolbar.#moveItemsBackToTheirOrigin
is not coupled to the existing overflowed item list, and instead takes
advantage of the insert-order provided when calling #overflowedInfo.entries()
and queries the DOM for the overflowed item IDs directly.

A later patch in this series is going to conditionally move overflowed
items into different panels based on whether or not they're provided
by WebExtensions, and so this change means that #moveItemsBackToTheirOrigin
is not going to care which panel the overflowed items are in, it should
behave the same.

Differential Revision: https://phabricator.services.mozilla.com/D160292

UltraBlame original commit: 3fbfc3d94e5aa181c9e0da0116b4f28e5533176d