Bug 1793803 - Add skipPresenceVisibilityCheck nsIClickRule option. r=timhuang
This allows us to disable the visibility check for certain rules - edge cases
where the cookie banner is hidden initially and the MutationObserver in
CookieBannerChild does not pick up on visibility changes.
This can happen because we are in an iframe and the frame itself changes in
visibility.

Differential Revision: https://phabricator.services.mozilla.com/D160649

UltraBlame original commit: 5c1ba6aae6af61fca03d39769e0219fbf2c85d13