Bug 1797688 - Part 9: Make gXPCOMThreadsShutDown DEBUG only. r=xpcom-reviewers,mccr8
`gXPCOMThreadsShutDown` is needed for the assertion in `ThreadEventTarget::Dispatch` but we do not want other shutdown checks to rely on it, as it is too specific for this case. If we ever would really need this checkpoint in time during runtime in release we should consider to make it become a new `ShutdownPhase` in between.

We move the state to `ThreadEventTarget` and have a `DEBUG` only method to prime that assertion over there.

Differential Revision: https://phabricator.services.mozilla.com/D160627

UltraBlame original commit: 9be2b7c981dd9e8181dc5b513c652ab7b6501fe6