Bug 1767055 - Support Glean metrics in Artifact Builds r=nalexander,janerik
If there's no `COMPILE_ENVIRONMENT`, generate jogfile.json during build.
It contains information about all the metrics and pings.

If we are both not MOZILLA_OFFICIAL _and_ mozilla::IsDevelopmentBuild,
we look for the jogfile on the first use of the Glean or GleanPings global.

If we find the jogfile, we load our metric and ping info from there,
making it available to JS.

JOG isn't instrumented as it only does anything when !MOZILLA_OFFICIAL,
which is exactly the condition for disabling ping upload (see rust feature
`glean_disable_upload`), so we wouldn't expect to ever hear anything.

Differential Revision: https://phabricator.services.mozilla.com/D160540

UltraBlame original commit: 021192aa8ef011e76effb69b3d3531b8b6ce1946