Bug 1798953 part 1: Initialize ClusterIterator::mIterator in the init list. r=jfkthame
Before this patch, we implicitly invoke the default constructor for this
member-var, and then immediately stomp on the results with a reassignment.

This patch avoids that unnecessary usage of the default constructor, saving
ourslves a bit of work and getting this member-var more directly into a
fully-initialized/valid state.

Differential Revision: https://phabricator.services.mozilla.com/D159903

UltraBlame original commit: 3ca372ab2ceca82327882e044a4bbcf42c65641f