Bug 1754823 - Add a method for expanding the perf categories. r=perftest-reviewers,AlexandruIonescu
This patch adds a method for expanding the categories. This method takes all the configuration details from the PerfParser (variants, apps, platforms, etc.) and combines them to create the categories that people will be seeing and selecting from. It creates the queries, and checks for any restrictions among them to provide selections that work. See the method for more details on all of this.

Depends on D160413

Differential Revision: https://phabricator.services.mozilla.com/D160414

UltraBlame original commit: 6e94007a5fe8a982adbad345bbb35afc14974cd1