Bug 1754823 - Add methods for getting tasks for the perf selector. r=perftest-reviewers,AlexandruIonescu
This patch adds two methods for getting/selecting tasks. The first one (get_tasks) is used to either query fzf directly, or to use interactively for user selections. The second (get_perf_tasks) makes use of the first to query for categories (through the user) and then query for the selected tasks. It also performs all the necesary intersections/unions for the queries.

These two are separated because we need a simple way to allow the user to look and select the raw tasks instead of categories if needed. This way the user can still get a comparison link at the end instead of going to `mcah try fuzzy`. See the next patches.

Depends on D160414

Differential Revision: https://phabricator.services.mozilla.com/D160415

UltraBlame original commit: 3c18a3f66f89685a90c6ef0151661ef67cc8ec6e