Bug 1790872 - Part 3: Add a unit test for the IPC sent/recvd message telemetry, r=ipc-reviewers,mccr8
This is done using the gtest framework, and adds a new protocol which has its
messages counted. As the IPDLUnitTest process doesn't count IPC messages with
glean, this only counts messages on the parent process side.

This test also checks that we don't count this information outside of nightly,
and that it isn't counted for messages sent within the same process.

Differential Revision: https://phabricator.services.mozilla.com/D158275

UltraBlame original commit: 4f712ee198549ced74ccb3c256aec9a30305f500