Bug 1799193 - Refactor XUL textbox-label/treeview-cell cropping to share code. r=emilio
This shouldn't change any behavior, it just refactors the code to use a shared string-cropping function.

It's still an approximation, and theoretically could be quite inaccurate in the case of complex text shaping,
but in practice is "good enough" for the kind of UI-element display where we use it.

Differential Revision: https://phabricator.services.mozilla.com/D161328

UltraBlame original commit: 8f567b0909e9f7adffe4856ab1cc154cae754020