Bug 1799947 - Update existing tests to put browser_action's into the navbar by default. r=willdurand,extension-reviewers
We had a number of tests that assumed that when adding a browser_action without
specifying the default_area, that the button would enter the navbar. The previous
patch in this series changes that assumption when the Unified Extensions UI is
enabled.

Instead of updating all of these tests to add additional steps to move the
browser_action's out to the navbar after adding them, I've gone ahead and
updated them to default their browser_action's to the navbar instead.

Differential Revision: https://phabricator.services.mozilla.com/D161721

UltraBlame original commit: ff58300200c458ce99963eda14a3716685b90b6d