Bug 1800225 - Only `realpath` the `prefix` when resolving the sysconfig packages path r=firefox-build-system-reviewers,glandium
On POSIX platforms there can be `lib` and `lib64` for `site-packages`.
On Fedora35 once a virtualenv is created, `lib64` is a symlink of `lib`.
Since we can determine what these paths will be before the virtualenv is
created, `realpath` has different results before/after, which is
problematic for our `pthfile`.

Since we only care about resolving the symlink in the prefix (to keep it
in line with the `topobjdir`, we can simply only `realpath` that and
disregard that there's a symlink between `lib` and `lib64`, which makes
our `pthfile` that's created and expected always the same on all
platforms.

Differential Revision: https://phabricator.services.mozilla.com/D162136

UltraBlame original commit: 79c3362f76c45101169b042db7d84d55fc30faff