Bug 1800462 - Revert the Windows part of bug 1791476. r=firefox-build-system-reviewers,andi
Considering the difficulty to figure out what's going wrong because none
of members of the build team can reproduce this, and considering that
the problem bug 1791476 addresses only affects TSAN for the moment, and
TSAN is not supported on Windows, let's just go back to what the linker
wrapper was before bug 1791476 on Windows, which will at least unblock
people.

We'll eventually have to figure out what's going wrong, but that's for
another day.

Differential Revision: https://phabricator.services.mozilla.com/D162179

UltraBlame original commit: 659083b7858bd86952ce52b157fd6ba200fda454