Bug 1461610. In image documents change the cursor to zoom in/out based on full zoom. r=emilio
Clicking on the image toggles between "shrink to fit" and full size mode. However it also
resets the full zoom. So if the full zoom has been changed then a click can zoom in/out.

The full situation is complicated because you could be in "shrink to fit" mode (so the cursor
should be the zoom in cursor) and then increase the full zoom so that image is bigger than it
would be in fullsize mode so clicking would actually result in zooming out. I don't try to
resolve this. Instead I just let cursor be decided by the "shrink to fit"/full size state if
that's relevant, otherwise the fullzoom will determine the cursor.

Differential Revision: https://phabricator.services.mozilla.com/D161980

UltraBlame original commit: 7e750a604f4ef65f0e83610f9c01ad56e89174ed