Bug 1800496 - Re-introduce channel tracking for cancellation; r=asuth
This was removed as part of an attempt to simplify worker cancellaton. However, this channel cancellation is important, as without it the timing changes when we finish a request. Without it, we continue to load a script, even after a worker has been shut down.

This tracking will help us reimplement cancellation.

Differential Revision: https://phabricator.services.mozilla.com/D162215

UltraBlame original commit: 6118d850a96dfa307d91cb73392c659ffb328335