Bug 1800496 - Re-introduce ArrayView; r=asuth
ArrayView is a data structure I removed as part of the simplification of worker cancellation, as it was only used for this purpose in the entire code base. I am bringing it back now, as we are re-implementing that same behavior.

Differential Revision: https://phabricator.services.mozilla.com/D162218

UltraBlame original commit: 03e408b6ebbf0880c8598a45d4e0c13941a7dd27