Bug 1801321 - Hide panel when pinning an extension to the toolbar.  r=mconley
Similar to the Manage/Remove/Report actions in the (context) menu, we
close the unified extensions panel when clicking "Pin to toolbar". This
will also prevent the panel to be empty when there is only one extension
listed and we decide to pin it to the toolbar.

Differential Revision: https://phabricator.services.mozilla.com/D162414

UltraBlame original commit: 655a5b20af885208712e79d278b55ddd849ddd0b