Bug 1788800 - fix new colorways card intermittent r=sfoster
From what I can tell switching to `withFirefoxView` and opening a new window to run these tests in + closing that window before disabling the theme messed with the timing of these tests in such a way that we started getting that vsync error shown in the logs. I've switched to running all these tests in the same window, which looks good from a couple of try pushes:

https://treeherder.mozilla.org/jobs?repo=try&revision=9a5f379ecc48bda08eab67210278b547f558f856

https://treeherder.mozilla.org/jobs?repo=try&revision=76b909924455fda54165ffeaa00b3654b4decd09

I'll file a follow up bug to change the default behaviour of `withFirefoxView` is to run tests in the same window, as this has come up before.

Also worth noting that even if this patch lands and fixes the new frequent intermittent we will likely still see the less frequent intermittents that happened prior to the change for Bug 1780441.

Differential Revision: https://phabricator.services.mozilla.com/D162617

UltraBlame original commit: 04f05185596d6331d91c7d9eb225628c1ada8c79