Bug 1802271 - Bootstrap the same base packages on all Linux distros. r=firefox-build-system-reviewers,ahochheiden
Nowadays, we don't need to bootstrap as many packages as we used to,
thanks to the bootstrapped tools and sysroots. The first step here is to
install the common base of necessary packages. As a matter of fact, none
of the other packages are necessary anymore, but we'll clean that up in
a followup patch.

This change and the followups have been tested on base docker images of
Arch Linux, Debian 10, 11, 12, Fedora 36, 37, Gentoo, OpenSUSE Leap,
Tumbleweed, Rocky Linux 8, Ubuntu 22.04, 22.10, Void Linux, and a Live
ISO of Solus Linux, with the assumption that actual users would have at
least what those images contain (and some of them are pretty barebones,
to the point that some of the base packages listed seem would seem
ridiculous... in fact we hadn't needed to include them so far).

Differential Revision: https://phabricator.services.mozilla.com/D162955

UltraBlame original commit: e2d7ecbf0690ddd7fb578063c9e43a8f4a95e03a