Bug 1802374 - Open unified extensions panel with keyboard navigation. r=rpl
It looks like this used to be possible by chance when we didn't check
the event in `togglePanel()`. Now that we do this (to prevent
context-click to open the panel for example), we need to check the event
type and which button or key is clicked/pressed.

Differential Revision: https://phabricator.services.mozilla.com/D163076

UltraBlame original commit: 579c9323bd7683eae42655d817f156701ff50a0a