Bug 1778510: Add a CallerType to ReduceTimePrecision functions r=timhuang
Along the way we remove the calls to JS::SetTimeResolutionUsec.
This function is used to set values that are only used when
we don't have a ReuceTimerPrecision callback, which we always do
(as confirmed by coverage.)  This in turns lets us remove some
other code.

Depends on D151298

Differential Revision: https://phabricator.services.mozilla.com/D151299

UltraBlame original commit: dba220581d933ba2452db6d1e9519ff695c1073d