Bug 1778510: Add RTPCallerType to GlobalObject r=asuth
This centralizes the logic in one place.

In order to do this, we will need to check the principal
off-main thread. (Well, we need to know if it's System
Principal.)  Worker and Worklet need special ways to do
this, so create a virtual method for it and let them
override it.  This is analogous to the ShouldRFP method
on GlobalObject.

Differential Revision: https://phabricator.services.mozilla.com/D157564

UltraBlame original commit: a8e623b228aeb04c741bebeeaea4f2524b2ddcc7