Bug 1219128 - Initialize Object and Function constructors when creating globals. r=jandem
Object and Function prototype initialization is stateful, and mutate the global
as it is initializing the content of the Global. Attempts at making this process
transactional was not successful.

Thus, `JSProto_Object` and `JSProto_Function` are now initialized as part of the
Global creation.

Failures to initialize the global might lead to have a Realm with a partially
initialized global. Such realm appears to have a valid global using
`hasLiveGlobal()`. When catching failures, such global might still be used
despite being partially initialized. A new field `initializingGlobal_` is added
to `Realm` to track whether the global is live or not. Failures to comply with
this requirement would cause the garbage collector to reclaim the Realm and
global in case of failure. This should be avoided during the initialization as
the global is rooted by the `GlobalObject::new_` function.

Differential Revision: https://phabricator.services.mozilla.com/D163016

UltraBlame original commit: 8ccf33e282b4632bcba7d9912e1db8ab1ac72ce7