Bug 1799692 - Part 1: Introduce ThreadsafeContentParentHandle, r=asuth
This introduces a new type to ContentParent which acts as a weak handle to the
actor and is safe to hold and manipulate from any thread.

This replaces accesses of the `ContentParent` type from the background thread,
as they were error-prone due to ContentParent not being threadsafe-refcounted.

The bulk of this patch is piping the new type through to the places it is
required, and removing now-unecessary extra complexity.

Differential Revision: https://phabricator.services.mozilla.com/D162346

UltraBlame original commit: 56bdce9421f30b02578b5e7f51c0f35438b714b6