Bug 1791754: Remove bogus assertion in gtk NativeKeyBindings.cpp. r=masayuki
As discussed in the bug, we're not justified in expecting this assertion's
condition to hold up; it can be made to fail in mundane/unconcerning ways.

DONTBUILD because this doesn't affect behavior at all in opt builds, and it
doesn't affect debug build behavior in treeherder-relevant ways either.

Differential Revision: https://phabricator.services.mozilla.com/D163416

UltraBlame original commit: 96ad3c23ee1b01907e0d59845616e3547b1b0346