Bug 1797284 - Teach about:logging to start the Firefox Profiler, with the appropriate profiler preset for a particular logging preset. r=julienw,mossop,flod
This has some provision to continue working if the tab is closed or reloaded,
but it's not fool proof. Eventually we might want to move this to a service, but
it's already very useful as it is.

Differential Revision: https://phabricator.services.mozilla.com/D160213

UltraBlame original commit: d4be9af4e5ae8c011f87a2770fe08a206d784fdf