Bug 1793463 - Part 7: Use dynamic registration for protocols handled by nsGIOProtocolHandler, r=necko-reviewers,valentin
This swaps out the previous code which hard-coded nsGIOProtocolHandler
handling inside of the IOService, and replaces it with
nsGIOProtocolHander dynamically registering and un-registering protocol
handlers at runtime in response to the relevant preference, so that the
protocols can be handled with the normal dynamic handler codepath.

The service is configured to start-up during 'xpcom-startup' through the
category manager, so attempts to use GIO protocols before then will
fail. If this ends up being a problem, we can move it earlier during
startup.

Differential Revision: https://phabricator.services.mozilla.com/D162806

UltraBlame original commit: 737e4828685ff1a061a4e2def7aee4c70439ee85