Bug 1782087, part 2 - Platform changes to use the doorhanger, r=timhuang
This patch integrates the doorhanger into the CreateCredential promise chain,
and in doing so added a few new helper functions PromptUserWithPolicy and FetchMetadata.

FetchMetadata is almost identical to the other Fetch* functions in this file.
PromptUserWithPolicy is a sllight variation on the other doorhanger-creating funcitons.

I updated the tests to pass by adding a new endpoint with the right headers and a static json file

Differential Revision: https://phabricator.services.mozilla.com/D162127

UltraBlame original commit: 955c8cad81122b2fee1e2e219bac921f1ace0983