Bug 1803171 - FedCM Promise rejections on timeout cause tab crashes, r=pbz
I found the source of the crash- it is how we handle the promise and timer pointers in IdentityCredential::DiscoverFromExternalSource.
The promise is easy. The timer is hard.

The right way to handle the timer pointer is to put it on the document (will be resolved by Bug 1803245).

I'm fine not having the timeout functionality for now (it is non-essential imo).

Differential Revision: https://phabricator.services.mozilla.com/D163364

UltraBlame original commit: 021d5430a274ff496c09037b7056b951652661d9