Bug 1801664 - Make <object> and <embed> behave like <iframe>. r=peterv
This flips the prefs:

* browser.opaqueResponseBlocking.syntheticBrowsingContext
* browser.opaqueResponseBlocking.syntheticBrowsingContext.filter

that makes <object> and <embed> behave more like <iframe> when it
comes to image loading. This landed on Nightly through bug 1595491,
but with this we let it ride to Beta and then Release.

Differential Revision: https://phabricator.services.mozilla.com/D163317

UltraBlame original commit: fe1f1bd756f3c3260ca80c7ec155d67a5aaa878f