Bug 1782573 - Verify JSContext is not used to alloc memory from helper threads r=nbp
The off-thread-inner-fcn.js test verifies that ParseContext is using
ErrorContext as AllocPolicy. It seems like a fragile test, so perhaps I should
convert it to a jsapi test.

TempAllocPolicy::assertNotJSContextOnHelperThread() needs to be defined in the
.cpp file because the header would have a circular dependency on JSContext.h if
it #included it. We do want to make sure it is inlined, though, to avoid the
extra function calls.

Differential Revision: https://phabricator.services.mozilla.com/D160697

UltraBlame original commit: 8e5880a50a25b4f3b4fa2426ed1036375d6ac8d6