Bug 1804340 - Generate path vertexes into pre-allocated output buffer. r=jrmuizel
This pre-allocates a vertex output buffer in DrawTargetWebgl so that we can generate
wpf-gpu-raster and aa-stroke output into it. This way, they don't have to realloc
a Vec for pushes or changing into a boxed slice. This can net 5-10% on profiles for
the demos noted in the bug.

Depends on D163989

Differential Revision: https://phabricator.services.mozilla.com/D163990

UltraBlame original commit: 99052dd249cc79a752aeed699b8cb4633f3d2cd9