Bug 1755564 - Tweak spell checker underline color in dark mode. r=spohl
I think this is a slightly better color and saves me from actually
making a decision over the options in comment 9 and so on.

I'll ask Firefox UX to take a look at comment 9 so that they can
evaluate and make a decision (and maybe tweaking our link colors), if we
want to change it.

Depends on D163402

Differential Revision: https://phabricator.services.mozilla.com/D163403

UltraBlame original commit: 05e6afe9862e0d569342f2426593348a58612c27