Bug 1703040 - Upgrade !mIsPositioned assertion to diagnostic assert. r=masayuki
We are currently winding up with crashes that seem to be caused by ranges in a
selection having start and end containers in disconnected trees. Most likely
this has something to do with the ranges crossing shadow or native anonymous
boundaries, but it isn't clear.

This patch changes `DoSetRange` so we should hopefully assert as soon as we
get into that situation, rather than when we try to inspect the range. The
most obvious way to do this check would be to upgrade
`AssertIfMismatchRootAndRangeBoundaries` from a debug assertion to a
diagnostic assertion. Unfortunately, those checks are rather expensive, and
this code should be fairly hot.

Fortunately, the existing `!mIsPositioned` assertion should have the same
effect in all of the situations I've been able to think of. I've re-checked
the code a few times since I wrote the patch, and I believe it is sound. We
should hopefully find out within a couple of weeks if we need to do more.

Differential Revision: https://phabricator.services.mozilla.com/D162242

UltraBlame original commit: c746a700bc16329b23fcaef760f48bee73584952