Bug 1803800 - load browserPlacesViews.js only once and immediately from browser scripts, r=mak,florian,Standard8
Unfortunately lazy script getters load the file once for each accessed symbol, which
redeclares "let" and "const" and "class" variables, which was a problem here. But also,
loading the file multiple times is bad for performance (it is not a module so we
actually reload it) and the "lazy" variables get dereferenced immediately from
markup that is present both in browser.xhtml and the hidden window on macOS, so I
doubt lazy loading was getting us anything performance-wise.

This patch simplifies things by 'just' loading the file from the markup.

Differential Revision: https://phabricator.services.mozilla.com/D163934

UltraBlame original commit: eabefc1c5d67ae50622e4617f7536f48779a22f0