Bug 1795687 - disallow opening content prefs in shutdown, r=mak
This prevents the issue described in the bug in two ways:
- avoid attempting to start a connection in shutdown
- call connection.close() even if we fail to add a shutdown blocker.

It does both because if we did only the first, we would still have a potential race where
when _getConnection is called, we would not be in shutdown, but by the time
we try to close the connection, we could be. If we only did the second, we'd keep retrying
which is also bad.

Differential Revision: https://phabricator.services.mozilla.com/D163225

UltraBlame original commit: 27da6226c2a0452d2e85d31c6b9b3e4d1205478d