Bug 1805430 part 1 - Handle dictionary shapes in Shape::replaceShape too. r=jonco
This makes the code for dictionary shapes more similar to the other shapes and
simplifies the callers.

It's also more efficient because instead of cloning the shape and then mutating it,
the new shape now has the right contents from the start.

Differential Revision: https://phabricator.services.mozilla.com/D164571

UltraBlame original commit: 6116446bee0d7040759293cdee089fc45ba371e4