Bug 1805199 part 2 - Support more getters on the global in CacheIR. r=iain
Before this patch, we supported only native getters that can take the global object
as `this` argument (DOM getters) for `GetProp` on `WindowProxy` and `GetGName`.

This patch extends this to also support getters (native and scripted) that need
the `WindowProxy` if we have one.

Differential Revision: https://phabricator.services.mozilla.com/D164541

UltraBlame original commit: 10ffcc14a7dac93a62b69938ac682f3eeff1652b