Bug 1805678 - clean up cruft in browserPlacesViews, r=mak
`.footer` is no longer used on extraClasses, so I got rid of that.

I simplified the `.entry` bits because they were unnecessarily
verbose. Honestly I think we may be able to remove the getter/setter
for `options` entirely, but I got bored of trying to modernize things...

PlacesPanelMenuView was introduced in bug 963095 and then replaced
as part of Photon (with PlacesPanelview) and then I removed the only
callsite (that was already behind a pref) in bug 1354117, so that was
just dead code, AFAICT, so I removed it.

Differential Revision: https://phabricator.services.mozilla.com/D164757

UltraBlame original commit: 110abf97facf9bc67b82466147c463b6befe2980