Bug 1804399 - Do not check the profiling directory inside _is_extra_profiler_run and gracefully fail later r=perftest-reviewers,sparky
Previously we were checing if there is a profiling directory to decide that
it's an extra profiler run. But actually this is incorrect because it can still
be an extra profiler run with a failed output. So we skip checking that
directory while deciding if it's a profiler run, and add an additional check to
see if this step is failed.

Differential Revision: https://phabricator.services.mozilla.com/D164662

UltraBlame original commit: ed1bf9574d1fde158a8d72bf3462b5612c922e59