Bug 1801277 - Add a shim for microsoft virtual assistant. r=twisniewski,webcompat-reviewers
The patch adds a shim for fixing the microsoft virtual assistant. The
iframe of the virtual assistant will fail when accessing the indexedDB
because the domain of the virutal assistant is categorized as a tracker.
Accessing the indexedDB in this case will throw a security error which
leads to the site breakage.

The shim replaces the indexedDB with a fake interface that won't throw
an error when accessing it to fix the problem.

Differential Revision: https://phabricator.services.mozilla.com/D165166

UltraBlame original commit: 7285da60030b5cce9bcdc83d40a4e88c34287baf