Bug 1805798: Downgrade 'control reaches end of non-void function' to warning r=sfink
In regexp-parser.cc, `MayContainStrings` switches on all the values of an enum class without a default case. V8 obviously isn't building with this as an error. Instead of sending a patch upstream every time this happens, it seems cleanest to treat this as a warning, not an error, in irregexp code.

Differential Revision: https://phabricator.services.mozilla.com/D165086

UltraBlame original commit: 29f618682a6a17908ffc624b74488f8f74d82797