Bug 1806944 - Fix flakiness in image-set tests. r=layout-reviewers,tnikkel
image-set-rendering-helper.js waits for load to create the element that
has the background-image, so there's no guarantee that the image is
loaded before the screenshot is taken.

Set up the test in DOMContentLoaded, to guarantee that the load event is
blocked by the images.

Differential Revision: https://phabricator.services.mozilla.com/D165452

UltraBlame original commit: 6f6ac157ed754b015e2c21df1791b8940053de3c