Bug 1807780 - Use -ivfsoverlay instead of liblowercase. r=firefox-build-system-reviewers,ahochheiden
Because this involves affecting the MSVC + Windows SDK packs, we need to
do this for toolchain tasks as well as Firefox builds in a single pass.

The MSVC + Windows SDK packs are altered to keep the original case
instead of lowercasing everything (except .lib files), and contain
an overlay file that we now automatically use from the WINSYSROOT.

This requires adjusting some paths to match what the original case
is, as well as removing everything related to the use of liblowercase,
which conflicts with the use of the overlay file. People using
liblowercase locally will still have a working setup as long as they
don't set WINSYSROOT (it's new, so they won't have it), and don't have
an overlay file in there (which they only would  if they ran the new
pack_vs.py).

Differential Revision: https://phabricator.services.mozilla.com/D165596

UltraBlame original commit: 9d4e6a46f3e49038b36ad1a77409ac34d69158e0