Bug 1807892 - Keep accounting for scrollbar sizes on menulist popup intrinsic size. r=tnikkel,layout-reviewers
See the comment for why we can't just use scrollbar-gutter: stable,
which is what we'd want, ideally.

This is reproducible on Win10 and also Linux as long as:

 * Overlay scrollbars are disabled.
 * The scrollbar size is bigger than the menulist arrow size.

This used to be done by the chunk of code here:

 * https://hg.mozilla.org/integration/autoland/rev/8f61a444ce01d08e22efe35220fd171ca77fb10f#l7.82

Which changed the size of the menu (not just the popup) which then the
popup expanded to. So quite hacky over-all.

Differential Revision: https://phabricator.services.mozilla.com/D165669

UltraBlame original commit: 832901f88c039f47966d0986ee43c15a7c7c8d55