Bug 1807669 - Remove faulty check for poisoned cell contents r=sfink
This check is pretty sketchy as it depends on specific bit patterns occuring
poisoned GC cells. And it turns out that it's wrong anyway. Since we now use an
PoisonedObjectValue, if we look at the second word of such a value in a 32 bit
build this function will return false instead of true (since the object tag is
even).

Instead, just check the free list, but restrict this to GC marking so it
doesn't have too much impact. I wasn't able to detect and increase in time
taken to run jit tests with GC zeal 10 with this change.

This fixes the valgrind failure. If there is stil a problem with accessing
uninitialized cell contents it will show up somewhere else.

Differential Revision: https://phabricator.services.mozilla.com/D165903

UltraBlame original commit: da122a067957f662a70c3fc4a637a4b0ae5a111a