Bug 1808210 - Check ptr equality in MConstantProto::congruentTo r=iain
It looks like basically what's going on here is `VisibleValues::forget`
can fail to be found in `set_` because it says it doesn't match itself,
because its `receiverObject_` is null. I don't know if this is the
preferred fix, or if we should check pointer equality in
`ValueHasher::match`. One interesting thing to note is that in
`ValueNumberer::leader`, we explicitly check a node for conruency against
itself in order to detect if it has opted out of redundance elimination, so
it would be somewhat awkward for `ValueHasher::match` to not allow nodes to
return false for congruency against themselves, although I think everything
would still be correct.

Differential Revision: https://phabricator.services.mozilla.com/D165912

UltraBlame original commit: e0bd0bb80d0f979ae42e0e2a889e46274a1cecab