Bug 1808642 - Remove word with small katakana chars from 1798297-1 reftest, because newer YuGothic versions have a bit of kerning there. r=dholbert
The tests for 1798297 assume that the 'kern' feature will NOT affect spacing in YuGothic, only the 'palt' feature does.
But testing with new YuGothic versions (e.g. on Win11) indicates that a bit of kerning is present, affecting the
small-kana glyphs.

So to keep the testcase working as intended, we remove the word that has small-kana characters. That wasn't significant
to what the test is addressing, which remains just as valid with the rest of the content.

Differential Revision: https://phabricator.services.mozilla.com/D166185

UltraBlame original commit: 600d37f1553ee59f40dd268378ea7a50f395402a