Bug 1808741 - Patch lit's styleMap to not set the style attribute on first render r=mtigley
lit calls `element.setAttribute("style", "<our styles>")` on first render of an
element so that it will have the style property set in the markup if it is being
server-side rendered. We don't need that, and it causes CSP errors in
about:logins so this patches out the functionality.

https://github.com/lit/rfcs/issues/13

Differential Revision: https://phabricator.services.mozilla.com/D165240

UltraBlame original commit: 5d3793ddd16ed337c5b26d0072bde43dedfd29b9