Bug 1806727 - Add an option to easily build with release configuration. r=glandium
This doesn't build with release branding, but allows you to create a
nightly build that has all the settings that we have on release, which
will be useful for profiling.

I could use --enable-release with a bit of effort, if you prefer
(probably renaming the current version of --enable-release to something
else).

All-in-all, suggestions welcome on this approach.

Differential Revision: https://phabricator.services.mozilla.com/D165219

UltraBlame original commit: 781b133ca6612db579e249daa800d30662552187