Bug 1806117 - reset docState after selecting PiP toggle with a touch event. r=pip-reviewers,mhowell
In Bug 1758103, we stopped default behaviour after detecting a "touchstart" event when selecting the PiP toggle; this was to prevent a click event from running and affecting video playback when PiP initializes. However, in doing so, we failed to restart the doc state properties "isClickingToggle" and "clickedElement", which are responsible for determining where and when we suppress other events. This patch ensures that we actually restart those properties so that we can interact with other elements on the content window after opening PiP.

Differential Revision: https://phabricator.services.mozilla.com/D166244

UltraBlame original commit: 98432f752ba20430cccce225f0011f7877c4c175