Bug 1810077 - Let the focus fixup rule ride the trains. r=smaug
While the spec issue and PR are not closed, there seems to be rough
consensus on our approach and timing, and I believe the current behavior
is better than what we're shipping on release.

So perhaps we could let this ride the train, and do minor tweaks in the
future if needed?

Differential Revision: https://phabricator.services.mozilla.com/D166763

UltraBlame original commit: 27e88caf4ca44142b910a30f0132fa327e01694d