Bug 1809752 - Part 2: Assert if GetCurrentSerialEventTarget is called on a thread pool with no active TaskQueue, r=mccr8
This will help catch issues such as the one from bug 1806751, as we will assert
earlier in the process. This also aligns better with the existing documentation
of the method (which has been updated), which suggests that the method would
assert if called on a thread pool.

Differential Revision: https://phabricator.services.mozilla.com/D166604

UltraBlame original commit: 188ed220db1ae3bb6210bd25b17abdf7a24a60ef