Bug 1810496 - Part 3: Use PodCopy when stealing marking work r=sfink
Pick a point in the middle of the mark stack to steal from and adjust it if it
is the top word of a two word entry. Now that we won't tear entries we can
replace the copying loop. PodCopy performs better here than memcpy because it
inlines the copy for small sizes rather than calling out.

Differential Revision: https://phabricator.services.mozilla.com/D166940

UltraBlame original commit: 517e684474559ff60861257921b901efefe22478