Bug 1809875 - Recalculate outdated frecency periodically. r=daisuke
Every time recalc_frecency is set to 1, annotate that there's frecency to
recalculate. Periodically check if there's anything to recalculate and arm the
task if necessary.
Note it would be possible, but more expensive, to send a notification or
directly invoke a method from cpp. We want updates to be cheap, and the cheapest
option is to just set a boolean.

Differential Revision: https://phabricator.services.mozilla.com/D166666

UltraBlame original commit: e226ed35355cea5a48f3199c3af504f8498e0aca