Bug 1790569 - Design a minimal Glean-based crash ping r=gsvelto,chutten
This adds both the glean crash ping definition as well as the metrics to be
sent in the ping. These metrics are very simple; in the future more metric
information (such as crash locations and details) will be included as well.

Differential Revision: https://phabricator.services.mozilla.com/D166925

UltraBlame original commit: 681f78a95580801f5e7ff949cfce34f80a916850