Bug 1809657: Start next line break chunk from start of previous search when no breaks found. r=jfkthame
This means that we start from a known non-break and that we shouldn't be in any
danger of causing false breaks once Uniscribe gets to unprocessed characters.

This also makes the crash tests manual and debug only.
Manual because now that the win32k pref is default on and not dynamic the tests
will not run on try any more.
Debug only so that we don't include code in opt builds that is only for manual
tests.

Differential Revision: https://phabricator.services.mozilla.com/D167271

UltraBlame original commit: 3c7123ec7398ecde73de1815291e001fe1c5b87c