Bug 1787974 - Fixed problem where AddTimerInternal() could add a new timer in the wrong spot. r=smaug
The previous version used IndexOfFirstElementGt which is implemented
with a binary search. That doesn't work correctly because the mTimers
array that is being searched is maintained in sorted order
*only if the "cancelled" timers are skipped*. This revision replaces the
binary search with a linear search which can handle this without
problem. Given that the number of timers present at any one time is
fairly modest, a linear search is probably a more appropriate choice
anyway.

Also added a new function, VerifyTimerListConsistency, which is called
once per wake-up in the main loop and enabled only in DEBUG builds, that
checks mTimers for ordering and consistency.

Differential Revision: https://phabricator.services.mozilla.com/D164294

UltraBlame original commit: f1f071954863dd6fb4a2a3a05eaf2a7f32562630