Bug 1805502: Fix an incorrect test-case in rtp-extension-support.html r=jib
This test case was asserting that extmap for rid is present for a unicast
transceiver. While it would be acceptable to have this extmap in this case,
it is silly, since rid will not be used. Also, add a bug link to the ini file
for a different failing test case.

Differential Revision: https://phabricator.services.mozilla.com/D165099

UltraBlame original commit: 66233c1a1d11b4a281ae06c9a47b0b3e0c660efd