Bug 1615546 - Remove the optimizations field from profiler frame table r=julienw
This requires https://github.com/firefox-devtools/profiler/pull/4432 to be
landed and deployed. We bump the profile version in this patch.

Afaik we never used the optimizations field of the frame table in the profiler
frontend and this pfield was null for more than 3 years (since Bug 1614622).
We don't have any plans to make use of it any time soon. It's good to cleanup
this field so it doesn't clutter the profile data and we can reduce the profile
data size.

Differential Revision: https://phabricator.services.mozilla.com/D167290

UltraBlame original commit: a78559e1f67b1f3136adb44fd86d65a80088bc5b