Bug 1809464 - Fix `test_principal` fail when `alerts.useSystemBackend=true` for Windows. r=nalexander,lina
Windows native notifications are only available in the parent process. The existing mochitest runs in a content process, therefore checking `"mozilla.org/system-alerts-service;1" in Cc` fails. The test interprets this to mean that the XUL notification backend is enabled, which is not the case.

Since this test seems XUL specific in nature, this change modifies the test to set `alerts.useSystemBackend=false` and removes the now redundant check on test startup.

Depends on D167477

Differential Revision: https://phabricator.services.mozilla.com/D166442

UltraBlame original commit: 04e0b2762d9a826577e24dfd25a5043caff7282f