Bug 1812379 - Re-clone the benchmark repo if the remote origins don't match. r=perftest-reviewers,kshampur
This patch fixes an issue where updating the repository origin causes a failure because we dont' check for changes there. With this change, we now check the URL before performing an update to make sure we have the correct repo. Otherwise, we'll fail on the revision validation step.

Differential Revision: https://phabricator.services.mozilla.com/D167834

UltraBlame original commit: 8714793c4e8d4d36301f536e41acc93fe1560ca0