Bug 1764443 - [devtools] Handle "object" jsonMl tags. r=ochameau.
In JsonML, "object" tag can be used to be replaced by an actual object, that
may or may not be formatted.
e.g. `["object", { object: 42 }]`, should be replaced with `42` in the returned
JsonMl array.

In order to do that, we need to parse the result of `header` and `body` recursively
to produce the final JsonMl that will be consumed by the client.

On the client, this means that we need to create object fronts for any object
actor that might have been created, so we can use them later (e.g. when the user
expand them, call the `customFormatterBody` method.

We also need to handle plain object grips (i.e. not formatted), and display them
using our regular `ObjectInspector`.

Some tests are added to ensure this works as expected.

Differential Revision: https://phabricator.services.mozilla.com/D164220

UltraBlame original commit: 9f139efe4d4c72f54b8fe76555d5acbde40cd1b9