Bug 1814049 - [devtools] Extend variant `devtools-no-eft` from 2023-03-01 to 2023-09-01 r=jmaher
We have not removed the no-eft mode yet, so we want to keep the variant alive.
We will try to pay attention to potential breakage with this job during the next 6 months,
and if we see that it didn't detect any implementation issue, we might remove it when we
reach the next deadline.

Differential Revision: https://phabricator.services.mozilla.com/D168529

UltraBlame original commit: 6c697fbbbb340273c6cc7553a8b2388440601e24