Bug 1814318 - Make L10nCache.ensure() re-cache strings when excludeArgsFromCacheKey is set. r=dao
This is an error in how I implemented the `excludeArgsFromCacheKey` mechanism in
D167318. It can be fixed by always re-caching strings when
`excludeArgsFromCacheKey` is passed to `ensure()`. We could instead store the
argument values of currently cached strings and then re-cache only when
different argument values are passed to `ensure()`, but that would require a
deeper change and I don't think it's worth it.

Differential Revision: https://phabricator.services.mozilla.com/D168509

UltraBlame original commit: 6bc7d03059ade62c0b8c13dbe681f49ad45f6cf7