Bug 1813960 - Ensure that GetContentSize doesn't return a size under the window min size. r=smaug
Otherwise, if we end up with pref-size < min-size, we resize the content
to be the pref-size pixel size, but the window size clamping comes in,
and sets the window size to be bigger.

If we're lucky, the other dimension changes, and we trigger a size
update which fixes up things, but if we're not we end up with this bogus
state where the viewport frame has the wrong dimensions and the resize
never arrives.

Accounting for widget constraints fixes it.

Depends on D168461

Differential Revision: https://phabricator.services.mozilla.com/D168462

UltraBlame original commit: 88e051666f4532b4fecbb0bc196346818bfe5f3c