Bug 1814613 - [devtools] Avoid firing DOCUMENT_EVENT for previous background page when reloading an add-on. r=devtools-reviewers,nchevobbe
This can lead to various issues as we were getting two sets of DOCUMENT_EVENT's when reloading an add-on.

Hopefully, this will be simplier to handle once we move WebExtension to EFT
and WindowGlobalTarget will be unique per document *and* their instantiation is solely
controlled by the watcher codebase.
We might still have to be careful to ignore this spurious about:blank document.

Differential Revision: https://phabricator.services.mozilla.com/D168045

UltraBlame original commit: ba79613c14b780ccfa1daed7889e6bf6528006a5