Bug 1814688 - Ignore some events in TerminatorTelemetry r=chutten
The events in question do not have associated histograms, but are included in
the JSON payload that nsTerminator writes.

The test for this functionality has been updated to reflect real conditions and
correctly fails without the fix in this patch.

Differential Revision: https://phabricator.services.mozilla.com/D168706

UltraBlame original commit: c3c671c7421aee658a0dd78d77ab521506de817a