Bug 1814855 - [devtools] Better message layout on narrow console output. r=bomsy.
First, we set the max-width of the message location in percentage
instead of vw, as the console output inline-size can be different
than the viewport one (e.g. when it is in editor mode).
We then turn the media query we had to hide the timestamp into a
container query, again, to take into account editor mode.
Finally, we place the location on its own row when the output is
narrow so the message text itself can still be readable.

Differential Revision: https://phabricator.services.mozilla.com/D168778

UltraBlame original commit: 20b7bd329af341292d38549a7c6d057dfd56ec7c