Bug 1815736 - Fix locales in flatpak desktop file. r=stransky
Locales in the flatpak desktop files have the language and the country
separated by a hyphen, but it should be separated by underscore.

"-mac" in Japanese locale is also removed according to bug 418485.

Spanish translation is added for the action to open the profile manager.

Differential Revision: https://phabricator.services.mozilla.com/D169220

UltraBlame original commit: 3f3970903e88f19edf9f66940ce49354efe6bb47