Bug 1817042 - Add foreground / background tab telemetry for Picture-in-Picture. data-review=darcese, r=mhowell
Notably, we cannot use visibilityState or detect whether or not the window is occluded,
as we currently force the platform to keep any windows that are PiP'd as "visible" in
the platform layer in order to keep things like subtitles working properly. So this
really only measures the time in the foreground of tabs, but cannot tell if the user
foregrounds a completely separate application to occlude the originating window.

Differential Revision: https://phabricator.services.mozilla.com/D169978

UltraBlame original commit: 1cfe650e2c496d506f9fe75a5a47d9ef63162b8e