Bug 1809360 - Have this test check against the correct scroll position. r=settings-reviewers,Gijs
The test is checking whether the preferences page has ever been scrolled at
all for any reason, which means the scroll that happens when the setting under
test is focused is causing the failure. But we're really only interested in
what happens when the setting is toggled, not focused, so we should save the
scroll position right before the toggle happens and compare the final position
against that, not just always against zero.

Differential Revision: https://phabricator.services.mozilla.com/D170087

UltraBlame original commit: 33b3a1ed83ab77fb1b0582a2f2849fa6b7a10a5f