Bug 1812132 - wasm: Allow anyref, structref, arrayref, and (ref T) in ToWasmValue/ToJSValue. r=jseward
The minimal JS-API spec for the GC proposal allows these types to be used on the
JS-Wasm boundary. So far we have not supported this. This commit updates these
methods so that the types work. This commit does not add support to our JIT
fast paths for exit/entry, so any use of these types will be stuck in the
interpreter entry/exit.

Differential Revision: https://phabricator.services.mozilla.com/D170005

UltraBlame original commit: 4bedc72a2b775918b0cf3d3a333075407e49fb61