Bug 1817507 - Process lines from confvars.sh more carefully during msix repackaging. r=nalexander
browser/confvars.sh contains a comment with "MOZ_BRANDING_DIRECTORY" in it, so
"if key in line:" was True, even though the rest of the code to determine a
value failed silently.
Address the problem with multiple checks: Strip whitespace from either end of
the line, and skip comment lines. And, if value is not set, don't return yet.
This is more than strictly necessary to fix the immediate problem, but should
avoid similar bugs later.

Differential Revision: https://phabricator.services.mozilla.com/D170262

UltraBlame original commit: 86f16f663b11ae4296a25e7a2be1ea499c42f126