Bug 1817829 - Don't enable wgpu's renderdoc feature by default. r=webgpu-reviewers,supply-chain-reviewers,teoxoy
It is unclear that we want it enabled for all users and this way we don't have to audit the changes in renderdoc-sys.
We can and should revisit this in the future if we feel like it would be useful to ship with webgpu+renderdoc integration.

Differential Revision: https://phabricator.services.mozilla.com/D170492

UltraBlame original commit: a1a8ef4bbfe48489ae61695cbd22427b6ce2d36a