Bug 1816346 part 1: Fix the transform tree creation test in browser_test_simple_transform.js. r=nlapre
Previously, the test was retrieving an Accessible with a non-existent id and calling testAccessibleTree on it.
That meant we were calling testAccessibleTree with null, which is a no-op.
Now, we get the intended Accessible so that the test actually tests what it was supposed to test.
This is important to test a potential bug in the next patch.

Differential Revision: https://phabricator.services.mozilla.com/D170306

UltraBlame original commit: 22b32e106152dcda58ff5dfac0663519b67ddf1c