Bug 1809518 - Recreate JSOracleParent instance when Utility Process crashes r=farre
Top-level actor can only be bound to one process, we can't reuse the
existing one to bind to a different process.

So if the utility process crashes and we've created a new one as a
replacement, we need to have a new JSOracleParent for the connection,
using the old one won't work.

This patch creates a new one upon the destruction of the existing one
if the reason for the destruction is AbnormalShutdown.

Differential Revision: https://phabricator.services.mozilla.com/D168482

UltraBlame original commit: 5cc0653b20720710880d813d748e7c8b9e57a806