Bug 1809518 - Update test_fetch_event* related tests based on the status of the ORB r=farre
These tests involve a bunch of no-cors fetch requests which use
`file_CrossSiteXHR_server.sjs` for sending the response back.
Depends on the request's URL, the response can possibly be

```
  response.setHeader("Content-Type", "application/xml", false);
  response.write("<res>hello pass</res>\n");
```

which will make JS validator to run and blocks.

The fix here is to toggle the pref of ORB when running the tests,
so we test both with ORB and without ORB behaviors, with different
expectations.

Differential Revision: https://phabricator.services.mozilla.com/D169276

UltraBlame original commit: 3eb92dc5aac9b587959f774b95205d70406403fb