Bug 1817443 - tidy up parameter handling in openLinkIn, r=mossop
When splitting out the window/tab/current cases, not all of them need all
these individual arguments. But it's hard to reason about which ones are used
where/how when we rename them all the time, and passing them all
individually is more verbose as a result of the renaming, so let's just not do
that at all.

There are a few "odd ones out" - `private` isn't a valid local variable name,
and both it and some of the other items were easier to deal with by just
referencing directly on `params`.

Differential Revision: https://phabricator.services.mozilla.com/D170211

UltraBlame original commit: 2b39dee429482aabc4a6f2f9f11ac0677f649293