Bug 1483869 part 2 - Rewrite bound function implementation. r=iain
This patch rewrites `Function.prototype.bind`. Instead of using a self-hosted
implementation that uses JS lambda functions, use a separate `BoundFunctionObject`
type and `JSClass` with call/construct hooks.

Using a type other than `JSFunction` matches the spec (and other implementations)
better. It also better matches the `WrappedFunctionObject` implementation we use for
shadow realms. It also lets us remove some complexity from `JSFunction`.

One of the main reasons for doing this is that it uses less memory: the self-hosted
implementation allocated a `JSFunction`, a `CallObject`, and an `ArrayObject` (for bound
arguments). The new implementation allocates only a `BoundFunctionObject` when we're
binding <= 3 arguments (the common case).

This patch stack improves Speedmeter's React-TodoMVC by a few percent. This test
allocates a lot of bound functions but has relatively few calls, so this is likely
due to reduced GC pressure.

The new implementation has a few minor differences in various corner cases with `toSource`
and the debugger. Another patch in this stack changes a few tests to account for this.

Differential Revision: https://phabricator.services.mozilla.com/D170566

UltraBlame original commit: 634f0d65b4adaa0cab542e211489728afb07d9ce