Bug 1818726 part 1: Include the cache in PDocAccessible::ShowEvent. r=eeejay
Previously, the cache was sent in a separate IPDL call.
Even though we delayed the show event, there was a chance that a client would walk the subtree and find a lot of Accessibles with no data.
Because no event is fired for those changes inside the subtree, the client might not know it needs to update them.
This resulted in missing information in NVDA browse mode in some cases.
Now, the tree is always in sync with its cached data.

Differential Revision: https://phabricator.services.mozilla.com/D171046

UltraBlame original commit: 0854bd8c6db1ea0ec68135a56bd7700b56662d19