Bug 1818400 - Optimize primitive dependency checks in picture caching. r=gfx-reviewers,ahale
The main goal of this is to fix an implementation detail where the
WR code had to read every primitive in the tile even when checking
if a small sub-tile was valid (as the advance amounts of the
primitive dependency array vectors was stored in each primitive).

However, this patch itself is quite a significant optimization, it
improves displaylist_mutate by ~16%.

Instead of maintaining separate arrays for each dependency, use
a single byte array and use peek-poke to store these dependencies.

This simplifies the code for comparing dependencies, and makes the
traversal of sparse index buffers of the primitive array much faster.

Differential Revision: https://phabricator.services.mozilla.com/D170710

UltraBlame original commit: bd35a3c3f2679f910ee036309bea47e18e58b1f4