Bug 1817761 - Key telemetry differently when using OHTTP r=necko-reviewers,kershaw
For our initial experiment we'll be using the `dooh.cloudflared-dns.com`
domain for our resolver.
We probably should key things separately for OHTTP, but for now this
should be enough to capture the telemetry we are interested in.

Differential Revision: https://phabricator.services.mozilla.com/D171674

UltraBlame original commit: 87c5e6e6e22070530a8e11301358a03c3767ab1b