Bug 1819343 - Allow MigrationWizard initialization to be kicked off either on binding to the DOM or with script. r=kpatenio,settings-reviewers,Gijs
Up until now, we've used the connectedCallback to initialize the MigrationWizard. That's
been fine, except that it requires us to remove and then re-add the element to the DOM
if we want to "reset" it and start over.

This patch adds a method "requestState" to the MigrationWizard that kicks off the
initialization. Embedders can set the `auto-request-state` attribute on the
element if they're happy to just use the connectedCallback.

Finally, this adds an intrinsic width to the entire MigrationWizard element to reduce
flicker when transitioning between states.

Differential Revision: https://phabricator.services.mozilla.com/D171742

UltraBlame original commit: 9339270f610961e8911a5b90ee0fae0701aba73d