Bug 1818603 - Refactor the LanguageIdEngine r=gregtatum
Refactors the LanguageIdEngine in a way that has equivalent
functionality as before but will be easier to add in functionality
to mock the LanguageIdEngine in a subsequent patch.

It also reworks the structure of the LanguageIdEngine worker code to
more closely resemble the structure of the TranslationsEngine worker.

Differential Revision: https://phabricator.services.mozilla.com/D171293

UltraBlame original commit: 44b561534ff6c4ac8c74331857a48e3e849d9642