Bug 1816131 - Improve the style and html of about:logging for output checkboxes and preset dropdown r=Itiel
This uses the actual ids for the labels so that clicking the labels
actually work.
This also adds some flex magic so that the labels are aligned with their
checkboxes.
Finally this indents the file part so that the hierarchy is better seen.

Differential Revision: https://phabricator.services.mozilla.com/D169486

UltraBlame original commit: fcfe1d5d0311cef66a2da22e79ed955fdd9e1e7b