Bug 1821051 - Refactor rule limit counting r=rpl
As a preparation for quota enforcement for session rules and dynamic
rules, this refactors the quota enforcement, by introducing a single
helper (RuleQuotaCounter) that keeps track of the rule counts and
is responsible for enforcing quotas.

This patch is purely a refactor, and the only behavioral changes are the
strings in error messages.

The error messages of static rules were previously not covered by unit
tests; this patch also improves test coverage in that area.

Lastly, removed TODO from the static rule loading. When a static ruleset
were to exceed the quota, the desired behavior is to ignore it and try
the next ruleset (this is consistent with Chromium).

Differential Revision: https://phabricator.services.mozilla.com/D172250

UltraBlame original commit: 9c145f2d20ea83a2190828d55877f8b469956ddc