Bug 1607634 - Part 3: Specialize NotNull for SideVariant, r=glandium,ipc-reviewers,mccr8
The idea behind this specialization is to provide the members which would be on
SideVariant<..> on NotNull<SideVariant<..>>, similar to how pointer-like
methods are available on `NotNull<T>`. This makes the type more ergonomic to
use from callers.

In the next part, this will be used for non-nullable actor members in IPDL
structs and IPDL unions.

Differential Revision: https://phabricator.services.mozilla.com/D168885

UltraBlame original commit: 5812856d98ee60142ac598adfebbb71f770f4d85