Bug 1823842 - Explicitly only build the wrench library on android. r=gfx-reviewers,jnicol
When upgrading the NDK to r23, the wrench builds for android fail
because cargo apk starts adding flags to the cargo rustc call it does,
and that's not compatible with the wrench crate having both a lib and a
bin target.

As cargo apk only packages the lib in the apk, we can just be explicit
and build the library only.

Differential Revision: https://phabricator.services.mozilla.com/D173244

UltraBlame original commit: ea4b6f8be5642f0cbf79b2ae6c4e8359fa93fe40