Bug 1824158 - Make vClipMaskUv highp again. r=gfx-reviewers,lsalzman
In bug 1823411 prim_shared.glsl's vClipMaskUv was made mediump,
assuming it was safe to do so as it is an unnormalized texture
coordinate. This is, however, causing fuzziness in a test on Adreno
devices, so we are now reverting it to highp.

Differential Revision: https://phabricator.services.mozilla.com/D173468

UltraBlame original commit: 9dd24249fabe8af5c5b8941c3d064c985e903181