Bug 1787182 [Linux] Don't fire glxtest process unless we know we really want to run r=emilio
Right now we fire glxtest on every Firefox start, even if we going to update, restart or ping running remote instance.
When we're running on system with broken/unstable gfx drivers (drivers/glx freezes or crashes) every such action is delayed or coredumps are generated on systems.

In this patch we launch glx test proces later if we know we need it.

Depends on D168650

Differential Revision: https://phabricator.services.mozilla.com/D168651

UltraBlame original commit: bde24058d87cb9ffeb955551c334556ff2c422fd