Bug 1825667 - Replace AddStaticElement suppression with one for InitializeStaticHeaders. r=webdriver-reviewers,jgraham
Changes in inlining decision in clang >= 16 make it so that the stacks
that the mozilla::net::AddStaticElement exception is meant to catch
doesn't match anymore because mozilla::net::AddStaticElement is not in
the stack trace anymore. Use the caller, InitializeStaticHeaders,
instead.

Differential Revision: https://phabricator.services.mozilla.com/D174193

UltraBlame original commit: 417ca586623b643f5f7d6147112ccc003a0bcd25