Bug 1823412 - Notify main thread blocked DLL loads only once per DLL in WindowsDllServices. r=mhowell,gstoll
We currently only send notifications for successful loads and for denied
loads. Denied loads are related to our blocklist, and they include
redirected loads and blocked loads.

For successful and redirected loads, the DLL will end up as a
loaded module in the process, so we can expect no further attempts to
load it. For blocked loads however, the DLL will not be loaded in the
process, which implies that we may see additional attempts to load it.

In particular, bug 1823412 shows that sending a notification with every
blocked DLL load can lead to an infinite cycle of events, if the DLL has
a hook registered with SetWindowsHookExW on WH_GETMESSAGE. To
execute the notification code that gets dispatched to the main thread,
the main thread needs to call into PeekMessageW, which will result in a
new attempt at loading the same blocked hook DLL, thus generating a new
blocked DLL load event, with notification code to dispatch, etc.

We should only need one notification per blocked module anyway, since
this is only used to log information about blocked modules in
third-party module pings. So, in this patch, we keep track of the
blocked modules for which we have already sent a notification, to be
able to send only one notification per module when it is was loaded on
the main thread.

Differential Revision: https://phabricator.services.mozilla.com/D173960

UltraBlame original commit: d79bd97efa79c1816f766ace5b3f6f5ee14f5f1f