Bug 1824433: Handle TRRServiceChannel in ShouldResistFingerprinting r=timhuang,tschuster
If the CookieJarSettings is not implemented, return false.

Also, change the error behavior to also return false - we
don't need to immediately return true on error here, because
it may be the case that something else indicates we should not
Resist Fingerprinting (e.g. maybe the TRR connection is to an
exempted domain.)

Because we are no longer returning early, we need to check if
principal is null, as sometimes it might be now.

Differential Revision: https://phabricator.services.mozilla.com/D173564

UltraBlame original commit: 77de6765779111874d99bf9a4827f4d71f99d989