Bug 860312 - [devtools] Check deprecated list before calling a getter. r=devtools-reviewers,ochameau.
As the drawback of not having the safeGetter evaluated is quite minimal, we taking
a very simple route and only check the getter name against the list of deprecated
getter names.
We didn't have test for deprecated warnings, so one is added where we check such
messages are not emitted when expanding an object inspector.

Differential Revision: https://phabricator.services.mozilla.com/D174057

UltraBlame original commit: 1bedc0cbd0cb4fcabb457f8e84455948b2916066