Bug 1806756 - part 9: Rewrite IME state tests of text control reframing to make it possible to run in remote content r=m_kato
The remaining tests are not required to run in remote content except some of
the iframe tests.  However, iframe tests are not so important because it's
managed in remote content from point of view of parent process.  Therefore, I
don't port it in this bug.

Differential Revision: https://phabricator.services.mozilla.com/D171204

UltraBlame original commit: 7117a64d83271adf86a00b24b64e0a15c4ff28ee