Bug 1824725 - [devtools] Add a global key shortcut to toggle JavaScript tracing. r=devtools-reviewers,fluent-reviewers,flod,nchevobbe
This key will be trigerred even when devtools aren't focused.
We are willing to trigger this key especially when the page is focused.
But the shortcut will only do something when the debugger is opened.

This feature highlights that JS tracing may as well work solely based
on the WebConsole... we may want to followup to make this feature work
across console and debugger.

Differential Revision: https://phabricator.services.mozilla.com/D174097

UltraBlame original commit: 637eac8f066b5ad3252ef63c2bb559d38b58cd38