Bug 1827281 - Include more information in the AssertIsDead assertion message, r=xpcom-reviewers,mccr8
This adds 2 additional pieces of information: the name of the call-site
which created the `ThenValue`, and the error code returned from the
`nsIEventTarget::Dispatch` call, if it occurred.

This will provide more context for errors from AssertIsDead to allow
finding the code which is buggy (using the call-site), as well as the
situation where the behaviour is happening (as we'll be able to know if
the issue is caused by the Dispatch call failing).

Differential Revision: https://phabricator.services.mozilla.com/D175072

UltraBlame original commit: bb6a77bffb41aef7641ddbeee5d728dd1ff345cb