Bug 1828413 - Ensure menubar menus are non-fractional, and re-enable browser_mouse_enterleave_switch_tab on Linux.
What was going on is that the menubar height was fractional so the
clientX/Y coordinates were as well and we never passed this on
automation:

  https://searchfox.org/mozilla-central/rev/31f5847a4494b3646edabbdd7ea39cb88509afe2/testing/mochitest/tests/SimpleTest/EventUtils.js#1305

This padding matches Adwaita. The previous padding was meaningless
because the native theme overrode it.

MANUAL PUSH: Trivial fix CLOSED TREE

UltraBlame original commit: b00871bc37dda58458db4293503fc6aadb5ea9e9