Bug 1824772: part 4 - Cache jit hints in the process for eager baseline compilation r=iain,arai
Add a cache, based on a bloom filter, in the JIT Runtime that will exist for the process lifetime.  Whenever a script is baseline compiled, we add a hint to the cache based on the script filename+sourceStart values.  If we encounter this script again, during a navigation for example, then we can eagerly compile this script and skip the warmup.

Using a bloom filter does introduce false positives, so we try to maintain a false positivity rate of less than 1% and if we exceed this value we clear the cache.

Depends on D175523

Differential Revision: https://phabricator.services.mozilla.com/D175524

UltraBlame original commit: 9d119497d6d63883ec22cfd4bb502109e3f917fb