Bug 1823082 - GeckoView PDF Stream Adjustment r=geckoview-reviewers,jonalmeida,amejiamarmol
If an issue occurs when generating the PDF stream, we should be able to
send an error. The prior function (SendError()) was for stream readers
to use. The error state here can occur during writing.

Created a WriteError() option to set an error state that will cause
an IOException when reading to indicate the stream is in an inconclusive
state and to stop reading from it.

Differential Revision: https://phabricator.services.mozilla.com/D175918

UltraBlame original commit: 684aa38de8613b622aed62937d70e6661d21e9ca