Bug 1824910 - Bug 1829100 - [devtools] Decode folder name in Debugger Sources Tree. r=devtools-reviewers,bomsy.
We were decoding groups and files source items label,
but not folders.
We fix this and take the occasion to use decodeURI instead
of the non-standard unescape.
A test is added to ensure the source tree items are rendered
as expected.

Differential Revision: https://phabricator.services.mozilla.com/D176145

UltraBlame original commit: e9234a427470a56fed9b62a4fbd31b42a24e4623