Bug 1829159 - GeckoView Request External App r=geckoview-reviewers,amejiamarmol,calu
Adjusted the options on a WebResponse to include a way to request that
GeckoView embedders ask if users would like to view the file in another
app or to show the confirm download flow. (Default is false.)

Set requestExternalApp to true if the GeckoView embeder should prompt
the user to see the file in another app.

Set skipConfirmation to true if the GeckoView embeder does not need to
confirm the download. (Default is false.)

Differential Revision: https://phabricator.services.mozilla.com/D176179

UltraBlame original commit: c03b6a689180cb27f0c1b308b692b1367c1ed842