Bug 1672273 - Provide enums in Rust and C++ for labeled_* metrics with static labels r=janerik
First we have to replace the venerable metric hashmaps with metric `match`
statements (like events). This is because, by adding the enum to the Labeled<>
type, we've made it impossible to use HashMap (since its type requires a known
Value type `V`, and Labeled<Counter, $SomeEnumType> is incomplete).

Then we had to generate meaningful enums. The order of the values is important,
as I foresee a future in which we may wish to use the enum variant value
(its "discriminant") to index into the glean-core `labels` list.

The enums can't have naming collisions. In Rust this is easy, so long as we
stay away from reserved keywords (bug 1814767). In C++ we are using `enum
class` so the variants' identifiers are scoped. However, on my system at least,
this is still susceptible to collision with preprocessor defines. So we prefix
the variants with the letter `e`. So far so good.

C++ Labeled metric implementations is regrettably moved from Labeled.cpp to
Labeled.h, as the template specialization is no longer full.

No decision about JS is made or implied by its exclusion from this commit.
This includes support for runtime registration (via JOG).

As for Rust, this patch maintains the string-only Rust API advertised by the
SDK's `Labeled<U>` trait. No decision has been made about whether to support
the now-codegen'd Rust enum labels in FOG only through implementation against
FOG's `LabeledMetric<U, E>` or by extending support into the SDK itself.

The use of strings for labeled_* metrics is preserved mostly unchanged.

The transformation from enum back to string label is performed in Rust during
`enum_get`. This is an improvement over the current situation as the strings
in use are guaranteed to not be allocated. However, I suspect we can improve
even more by moving this translation lower (into the SDK, ideally).

GIFFT support requires getting the label string from the enum, which is exposed
on the FFI.

Differential Revision: https://phabricator.services.mozilla.com/D170109

UltraBlame original commit: b1508e3f22f59fdc769d434123800bd04dba14d2