Bug 1829049 - Explicitly use ~A or ~W variants of Windows functions. r=jandem
As of bug 1684110, we avoid using the generic names for Windows API
functions, and either define UNICODE or use the explicit variants ending
in ~A or ~W.

That didn't end up applying to js because it wasn't using STL wrapping so far.

As we're about to enable STL wrapping, we can either define UNICODE or
use ~A or ~W variants, but we choose the latter because some uses
involve more work to switch to the ~W variants. At least, it's not
explicit that the ~A variants are used in places.

Differential Revision: https://phabricator.services.mozilla.com/D175996

UltraBlame original commit: 15a316022a7a68fef9af39590bbbfe94679ecc94