Bug 1829896 - Part 1: Minor optimisation to allocation site updates r=jandem
This is a very minor optimisation so that we only read from a single location
when upating an allocaiton site rather than two. The allocation count being
zero implies that the site is not already on the list of active sites.

Differential Revision: https://phabricator.services.mozilla.com/D176436

UltraBlame original commit: 2c29263b7c19797ccc30e9a831f2086a98a42e95