Bug 1830068 - Optimize MLexicalCheck in FoldLoadsWithUnbox. r=iain
A common MIR pattern when loading lexicals from an outer scope is
`MLoad{Fixed,Dynamic}Slot => MLexicalCheck => MUnbox(fallible)`.

This patch makes it possible to eliminate the lexical check in this case and to
fold the load and unbox into `MLoad{Fixed,Dynamic}SlotAndUnbox`, because this
instruction will bail out if we see the uninitialized-lexical magic value so a
separate check for it is redundant. This is similar to how we don't need a separate
hole check for `MLoadElementAndUnbox`.

We have test coverage for this case in multiple jit-tests.

Differential Revision: https://phabricator.services.mozilla.com/D176549

UltraBlame original commit: e8232b1436b97333962c68790f3925500412fdb0