Bug 1826630 - wasm: Add OOM testing for GC object allocation. r=jseward
This is pretty simple, but found an issue with the trailer
recycling code where we need to report OOM's. We also need to
ensure that the trailers added and removed vectors stay in
sync in the event of an OOM and retry.

Differential Revision: https://phabricator.services.mozilla.com/D174790

UltraBlame original commit: 3f722a2895e67b5822b64a9f7c0a02585c3bcbee