Bug 1822624 - Add subsuite support to mozlog, r=ahal
A subsuite is intended to be a subset of tests within a suite that are
run with a specific configuration. For example it might be just the
rendering related tests run with an experimental graphics
backend. This is a common requirement when landing significant new
features or rewrites.

We assume a requirement to run multiple subsuites within the same top
level suite (e.g. to improve performance by running the subsuites in
parallel).

mozlog currently doesn't support this concept well because we assume
that there's only one suite running at a time, and attach run_info
configuration data to the suite itself.

To support this, we add one new log action: add_subsuite(name,
run_info=None). This allows specifying a subsuite that will occur in
the context of the current suite, and allows specifying the updates to
the base run_info data relevant to that suite (typically this should
involve at least one key of the form {"subsuite": name}).

We also add subsuite keys to other parts of the logging that are
associated with specific test configurations (i.e. the test_* actions,
and the actions that report the lsan data, asserts, etc. that may
correspond to a specific subsuite configuration).

Differential Revision: https://phabricator.services.mozilla.com/D171999

UltraBlame original commit: 01a92528ff1b7be2661876d4190c57a8890a86a6