Bug 1830921 - Discard JIT code when disabling or enabling the nursery r=jandem
This makes it so we always discard JIT code when the kinds of GC thing we can
allocate in a zone changes. Previously this happened in two separate places for
pretenuring, and didn't happen when we enabled/disabled the nursery.

Differential Revision: https://phabricator.services.mozilla.com/D177272

UltraBlame original commit: 04005f8322707e3dc71ee5b2115ad42d5a958a8d