Bug 1831126 - Fix mach try perf selection with chromium-as-release tasks. r=perftest-reviewers,sparky
Modify the selector so that the new `chromiums-as-release`/`custom-car` tasks are properly selected when a user submits a job. Previously custom-car was going through on firefox only selections. This patch should fix this as well as add a new --custom-car flag for selection.

Differential Revision: https://phabricator.services.mozilla.com/D177073

UltraBlame original commit: 095c03334b5f89857267cfb3b9c7243471ffbd3f