Bug 1824083 - Add API to allow Gecko to request a new Surface from the application. r=geckoview-reviewers,m_kato
Adds a new interface, GeckoDisplay.NewSurfaceProvider, of which an
implementation can be provided in the GeckoDisplay.SurfaceInfo passed
to GeckoDisplay.surfaceChanged(). We include an implementation in
the GeckoView class, which works by toggling the SurfaceView's
visibility, causing a new surfaceChanged callback to be fired.

Differential Revision: https://phabricator.services.mozilla.com/D176720

UltraBlame original commit: e26fa9885b0e296812a73a34b657f11694d29533