Bug 1831829 - Consider multicol container in paginated context when setting mIsTopOfPage bit. r=emilio
A top-level multicol in a paginated context can be fragmented. We should use the
same logic to set the `mIsTopOfPage` bit as if the multicol is a nested one.

WARNING: loading the testcase without this patch can trap Firefox in an infinite
loop.

Differential Revision: https://phabricator.services.mozilla.com/D177578

UltraBlame original commit: 1620189277fc5ccc3e0afdf9ddba9de9794491e4