Bug 1828511 - [devtools] Add custom formatter support in Scopes panel. r=devtools-reviewers,sebo,ochameau.
The scopes data we get in Scopes panel (from `FrameFront#getEnvironment`) don't
contain any object front, only grips.
As the setup is already complex, I didn't want to change that, so I went the
way we already do in some other places and passed the client to the ObjectInspector
we render; this way the CustomFormatter element can call `createObjectFront` from
it when it does not receive a proper front.
This revealed an issue in how we were passing the `front` prop from `renderJsonMl`,
as we weren't doing any check on it, and as a result it might have been a simple grip.
We also decrement the `autoExpandDepth` in `toggleBody` so we don't get auto expand
"regular" objects that might be inside custom formatter body.

Differential Revision: https://phabricator.services.mozilla.com/D175900

UltraBlame original commit: 4e98ea8e562ced819eb81f59f2a51eb740f5fe34