Bug 1827541 - Add the editor's legacy mode behavior test tasks to prevent new regressions r=jmaher,jgraham,taskgraph-reviewers,sylvestre
Until unshipping the legacy behavior of the builtin editor, we need to keep
testing them because currently Nightly and early beta builds test only the
new behavior so that regressions firstly appear in first late beta build,
thus, it may be hard to fix the regressions without risky patches.

This patch adds tasks to check the legacy behavior with mochitest-plain,
mochitest-chrome, mochitest-browser-chrome, mochitest-a11y, web-platform,
web-platform-reftest, and the crashtests.  It's enough to check it in
mozilla-central (i.e., check once or twice a day) because changes for the
editor module won't be landed so frequently and once the regression would be
found in the Nightly channel, we can consider fix it or back the regressors out.

Differential Revision: https://phabricator.services.mozilla.com/D175730

UltraBlame original commit: c132cd7a59209dd2f043514e253f0cc42f0d04ad