Bug 1827684 - Allow audio/aac. r=sefeng,necko-reviewers
It's possible for a streaming server to not start at an ADTS header
sync marker. We need a sliding window sniffer to be able to handle
this, until then we'll include audio/aac in the gradual ORB
transition.

Differential Revision: https://phabricator.services.mozilla.com/D178177

UltraBlame original commit: 495efbeb53a2e909f2ab5a5e1754b842dbdc0229