Bug 1817997 - Update the SourceBuffer and MediaSource implementation to use the new TimeUnit. r=alwu
Similarly to the previous patch, it's necessary to reduce the resolution of
values visible from script to keep the exact same behaviour as before. We might
be able to remove this eventually but it might not be very useful to do so, the
benefits of using rational-based time are mostly for internal computations.

Differential Revision: https://phabricator.services.mozilla.com/D173311

UltraBlame original commit: 8e0dcf16313894d34a5b26ec183f10e342ca434d