Bug 1833114 - Use separate mark stacks for black and gray marking r=sfink
This adds a separate auxililary mark stack and switches to it when necessary.
Switching is done by std::swap so that |stack| is always the stack for the
current mark color. We avoid swapping if both stacks are empty so that we don't
always use the auxiliary stack.

This removes a bunch of state from MarkStack and assertions around ensuring
that state is valid.

Differential Revision: https://phabricator.services.mozilla.com/D178040

UltraBlame original commit: 8b4f4da78faaea2ee6db5c656c6c0c286c9bbc03