Bug 1833624 part 4 - Call RegExp stubs also from Baseline ICs. r=iain
This is faster, especially for the `Matcher` and `ExecMatch` stubs, because allocating
the match result array is faster from JIT code than from C++.

Speedometer 3 shows some medium/high confidence improvements for various sub tests.

Moving the input registers into place was one of the tricky parts. After trying a few
approaches, using `MoveResolver` was simplest so that's what this patch implements.

Because the regexp stubs are held weakly by `JitRealm`, this patch stores a strong
reference to the stub's `JitCode` in the IC stub. It isn't actually used for the call,
because there are some edge cases (that we shouldn't hit in practice) where we discard
JIT code including the regexp stubs, but this doesn't work for scripts active on the stack.
To make sure we always call the correct regexp stub code, we load it from `JitRealm`.

Differential Revision: https://phabricator.services.mozilla.com/D178331

UltraBlame original commit: c19f93169200ad2c4910beda47347e46f512859f