Bug 1832585 - [devtools] Clear selected location in sources reducer on target removal. r=bomsy
The selected location holds a reference to a source and sometimes a sourceActor,
which may become irrelevant when navigating or when a target is removed.
Not doing this leads to editor being kept opened for sources which are now defunct.

Differential Revision: https://phabricator.services.mozilla.com/D178648

UltraBlame original commit: 78ee8f48414cd88dc5ca6146a90ab89b7c92f571