Bug 1829687 - Add a mechanism for deducing the preferred languages for a user; r=nordzilla
This is a big logic change, but with switching the views and adding the
translations command to the app menu, it's much easier to invoke the
translations panel in uncontrolled situations. This patch adds a
best-effort approach to deducing an appropriate language for the user.

Differential Revision: https://phabricator.services.mozilla.com/D179085

UltraBlame original commit: 6a44f24cc6be6517536cdbdab50526c3a28ebbc2