Bug 1833653 - Create mochitest-plain tasks on Wayland, r=jmaher,taskgraph-reviewers
This sets up a new worker-type and entries in `test-sets.yml` and
`test-platforms.yml` for a new pool running with Wayland enabled in a
virtualbox vm.

This test platform is a bit different, in that it uses generic worker and a
specialized pool. So we can't use the normal `instance-size` method of
selecting worker types, and instead rely on setting `worker-type` directly.

These tasks are not running anywhere yet, the primary purpose of this patch is
to make it easy to run them on try.

Differential Revision: https://phabricator.services.mozilla.com/D179364

UltraBlame original commit: 3b1d733c983500fb67921230cfbf07b202d19ec5