Bug 1810509 - Add a clickable invisible space around the UEI menu button. r=willdurand,extension-reviewers,desktop-theme-reviewers,dao
A unified extension item has two buttons: an action button and a menu
button. They are visually separated by 8px. The description string in
the action button depends on which of the buttons is hovered - the
action button, the menu button, or neither. If the mouse moves from the
action button to the menu button, it has to cross the 8px gap where
neither is hovered, causing the description to flash quickly between 3
strings. This patch circumvents that issue by making the 8px gap an
invisible padding of the menu button. This also removes the inert area
where you can move the mouse within the item and see a tooltip with the
name of the extension, but clicking would have no effect since a button
is not hovered.

Differential Revision: https://phabricator.services.mozilla.com/D177979

UltraBlame original commit: 37b731020cf74fb26ae5e16ff2f5252fe413cfb1