Bug 1835032 - [devtools] Always show number of occurences in All Issues section. r=devtools-reviewers,jdescottes.
We used to directly render the element using the property when
it was only used in one element.
For consistency, let's always show the X occurence details element,
and have it open if the property is only used in one element.

Differential Revision: https://phabricator.services.mozilla.com/D179224

UltraBlame original commit: 704d157b4af6110713a98568f41a09e168144879