Bug 1835717 - Disable _FORTIFY_SOURCE by setting it to 0 rather than undefining it. r=tjr
Clang 17 made a change that made it an error to undefine _FORTIFY_SOURCE
on mac, because it's defined as a builtin when ASAN is enabled.
Defining it as 0 has the same effect of disabling it, but avoids the
error.

Differential Revision: https://phabricator.services.mozilla.com/D179398

UltraBlame original commit: 8d79f3579b56da8de295914798ecec26ed8cb58d