Bug 1832898 - Add static labels support to runtime-registered metrics (JOG) r=perry.mcmanis
The initial implementation of JOG didn't care about static labels. I'm not sure
if that was an oversight on my part or what. But it mostly worked because in
most cases static labeled metrics work a lot like dynamic labeled metrics.

But then test_PingCentre.js began relying on the one behaviour that differs:
recording to a label not in the static list records to `__other__`
_without an invalid_label error_.

So now let's support it properly.

Note the use of ordered_labels here is only to get a list that's serializable.
The order doesn't actually matter so long as it's stable in the jogfile.

Differential Revision: https://phabricator.services.mozilla.com/D179239

UltraBlame original commit: 4d312f13db41056653bdfb69d955606feac41fc7