Bug 1831259 - prevent initialization re-entrancy for preference panes, r=mconley
This should fix the issue experienced by the reporter. It also pushes the
re-entrancy guard into the the 'init' method we define on gCategoryInits
objects, and removes the asynchronicity which was only there for fluent's sake.

Instead of blocking the insertion on fluent, which meant that for the in-page
find functionality we do N initializations and fluent passes, we make each of
the 2 consumers responsible for checking translation has completed. This means
find in page now just has 1 fluent pass, instead of N passes for N categories.
This should speed up the find in page initialization, and means initialization
of a category is now sync instead of async.

Differential Revision: https://phabricator.services.mozilla.com/D178232

UltraBlame original commit: af6ffefb5572c12356816adbe12d2b9887877ab0