Bug 1835559 - Ensure that text content survives translation r=application-update-reviewers,flod,fluent-reviewers,nalexander,settings-reviewers,Gijs
When an element is re-translated, the textContent is currently dropped.
To avoid that, make it an explicit part of the message.

This requires changing the l10n message ID.
Ideally, the message ID wouldn't change, but that is not possible until
the content can be preserved as a feature, i.e.
https://github.com/projectfluent/fluent.js/issues/169

Differential Revision: https://phabricator.services.mozilla.com/D179323

UltraBlame original commit: 81758aba6ab346d2a024fa3fbfacdbf5fbae0b62