Bug 590993 - Use new tab to load web handlers when in a pinned app tab r=Gijs
Unsets the browsing context for schemes (except web+ & ext+) in WebHandlerApp when launchWithURI is called in a pinned tab.

This will force the handler to load in a new tab instead of unloading the current pinned site, as long as the host of the web handler is different from the pinned site.

Differential Revision: https://phabricator.services.mozilla.com/D178880

UltraBlame original commit: 7f257a43b1b32958a25d847db07dd22dbe735505