Bug 1778275 - [devtools] Make selectSource helper select the precise source passed as argument. r=perftest-reviewers,devtools-reviewers,nchevobbe,sparky
(Instead of forcing to select original or generated based on shouldSelectOriginalLocation flag).

This also add an await for the selectLocation action in order to better wait for all async work.
This required to tweak the minified test as we don't want to wait for full selection.

Differential Revision: https://phabricator.services.mozilla.com/D179610

UltraBlame original commit: 00af548d93acc8c2c3ab652a04c882e3a80476a9