Bug 1836590 - Update standard font lists to support Ubuntu 22.04. r=emilio
Our original list of "standard" fonts on Ubuntu was based on the LTS release 20.04,
but there is now a newer LTS release 22.04 that has a substantially updated collection
of fonts. So for an improved user experience when anti-fingerprinting is in effect,
we should support the 22.04 font list.

Differential Revision: https://phabricator.services.mozilla.com/D179870

UltraBlame original commit: 07a9bab5ac94357f2a8efa4e676d57d3e13570da