Bug 1834038: Rewrite Ion implementation of ArrayPush to match CacheIR r=jandem
Prior to this patch, when reaching capacity, the Ion implementation of ArrayPush called ArrayPushDensePure to reallocate and store the value, but the CacheIR implementation called addDenseElementPure to reallocate, and then returned to jit code for the store. This could lead to a bailout loop, because ArrayPushDensePure is more strict about reallocating arrays with many holes than addDenseElementPure.

This patch rewrites the Ion implementation to use the same strategy as CacheIR. I considered sharing a masm implementation, but there are enough differences between the two cases to make it awkward.

Differential Revision: https://phabricator.services.mozilla.com/D179837

UltraBlame original commit: 79c414a997b5687eff2cf3ee319bc88ee9c68b9d