Bug 1834224 - Move JS_TRY macros to internal header. r=jandem
These macros are not usable by embedders because they expand to
JSContext method calls which are not public. However, they are used in
several places in the codebase, so we need to keep them somewhere.
JSContext.h seems like a good place for that.

Depends on D180556

Differential Revision: https://phabricator.services.mozilla.com/D180557

UltraBlame original commit: 8875630473bb428a95da292f5032214df3e26cda