Bug 1835180 [wpt PR 40241] - [GridNG] Prevent zero-sized set track counts at non-zero indices, a=testonly
Automatic update from web-platform-tests
[GridNG] Prevent zero-sized set track counts at non-zero indices

This CL fixes a a divide-by-zero in `ComputeTrackSizeRepeaterForRange`,
but the fix is nowhere near that code. The basic premise being violated
is that sets collections should always have the first set with a
`track_count` of 0, and no other indices should have a `track_count` of
zero. This is because the first set isn't actually a track
representation, it's for accumulated extra margin.

This premise gets violated when copying the set values from the parent
with opposite-direction children, which I added in
https://chromium-review.googlesource.com/c/chromium/src/+/4484336

The proper mitigation for this is to take the offset from the current
index, but the track count from the *next* index for opposite direction
subgrids. This will properly account for the initial entry with zero
tracks.

Bug: 1445565
Change-Id: I25586835034028f594895fa052d238a389383928
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4559196
Reviewed-by: Ethan Jimenez <ethavarmicrosoft.com>
Reviewed-by: Alison Maher <almahermicrosoft.com>
Commit-Queue: Kurt Catti-Schmidt <kschmimicrosoft.com>
Cr-Commit-Position: refs/heads/main{#1150006}

--

wpt-commits: fe0790161e433669646afe44fafa0972d4fc7d8b
wpt-pr: 40241

UltraBlame original commit: 246f996a3fe32081e68403aa14ed622a02316e3e