Bug 1837367 - Allow file migrators to show an error message if migration failed for some reason. r=kpatenio,desktop-theme-reviewers
The spec calls for a special-case in the event that file migration
ever fails - the wizard should be sent back to the selection page,
have the associated file migrator still set in the dropdown, and
show an error message.

A later patch in this series will add such an error message for
the bookmarks file migrator.

Depends on D180458

Differential Revision: https://phabricator.services.mozilla.com/D180490

UltraBlame original commit: 495fb1f0b39015786868e5486f2b752b820153c8