Bug 1838056 - Make sure to load MigrationUtils in BrowserGlue. r=niklas
Bug 1823537 moved the registration of the MigrationWizard JSWindowActor pair
into MigrationUtils from BrowserGlue. This makes it so that the wizard can
be used properly in the scenario where BrowserGlue hasn't yet had a chance
to run (which is the case when we're showing the migration wizard on startup).

This, however, means that the actors aren't being registered unless MigrationUtils
is loaded, which is no longer the case having moved the registration outside of
BrowserGlue.

This means that the embedded migration wizard in about:welcome (and the spotlight
experience in bug 1837051) will not work until that module loads and the
registration occurs.

Differential Revision: https://phabricator.services.mozilla.com/D180725

UltraBlame original commit: c065a7c4f8c0efdfa8a0fd0fcbac8da6a114e01c