Bug 1825777 - Improve formautofill heuristics by always identifying address fields r=mtigley,credential-management-reviewers DONTBUILD
At present, we don't execute address field identification for regions that don't support
address autofill. However, this impacts the accuracy of identifying credit card fields too.
This is due to the fact that if a field appears to be both an address and a credit card field,
we will consistently regard it as a credit card field if we don't execute address field heuristics.

This patch guarantees that we run both address and credit card heuristics when form autofill
is enabled for a region. We will still not autofill or capture address fields for regions
that do not support address autofill.

Depends on D179688

Differential Revision: https://phabricator.services.mozilla.com/D179689

UltraBlame original commit: bf687698c6425dfc6ca9e3ec388531063854a9e4