Bug 1838058 - Don't trim based on mp4 container duration when using MSE. r=alwu
Each mp4 fragment on huya.com has a duration, that is wrong, likely because of
rounding on the broadcaster side, cutting a short bit of audio at the end of
each segment. It's best to not trim the packet when using MSE to sidestep those
issues.

Differential Revision: https://phabricator.services.mozilla.com/D180839

UltraBlame original commit: 9891d17d6abe2c190871d5ca239c67797208a6e1