Bug 1832854 - pt3 - new vendor_and_commit.py script. r=ng DONTBUILD
This vendors a commit, handles the add/delete/rename mercurial
operations, and uses a provided commit message when committing.

Note: this is a python adaptation of a large bit of functionality from
fast-forward-libwebrtc.sh so that we can reuse the same logic between
fast-forward-libwebrtc.sh and an upcoming script to help with doing
cherry-picks from upstream in a standardized way.

Differential Revision: https://phabricator.services.mozilla.com/D180114

UltraBlame original commit: e1baa082edce77dcc9998daa11de3d636551d0ed