Bug 1837305 - Part 2: Add ShapeType for BasicShape parser. r=emilio
For outline shapes, fill-rule should be ignored. We add the flag in
the parser of BasicShape, so offset-path can use this flag to ignore
fill-rule.

Note: "ShapeType" means this property uses filled shapes or outline
shapes. For outline shapes, we ignore fill-rule. This is from the
concept of `<outline-shape>` and `<filled-shape>` in
https://github.com/w3c/csswg-drafts/issues/3468#issue-393916361

No behavir change in this patch, just add the ability for offset-path to
ignore `<fill-rule>` when combining all basic shapes into offset-path.

Differential Revision: https://phabricator.services.mozilla.com/D179625

UltraBlame original commit: 6f345b692ecbd7a6de126296efb8811b6bd8597d