Bug 1837467 - StartForceKillTimer only after SendShutdown has been called. r=smaug
The parent process might be busy such that the time that passes between NotifyTabDestroying and ShutdownProcess can become significant in terms of activating the ShutdownKill timer too early. In order to avoid this, we will start the timer only after we actually sent the shutdown message.

To cover the time that passes between NotifyTabDestroying (right after sending a delete to the Browser actor) and the effective NotifyTabDestroyed (in response to the Browser actor destroy roundtrip through the content process), we add a second timer that triggers the sending of the shutdown message.

Differential Revision: https://phabricator.services.mozilla.com/D180959

UltraBlame original commit: cd1285d0e9e6a63ba48a44b869c26a2252e71f5e