Bug 1802193 Part 6: Make browser_fullscreen-tab-close-race work with asynchronous fullscreen transitions. r=edgar
This ensures that the calls to SpecialPowers.spawn are always awaited. It
also changes setupFn to setupAndCompleteFn to clarify that that function
returns a promise, which needs to be setup before the click and awaited
afterwards. It also removes the "async" keyword from functions that return
Promises, but don't await them.

Differential Revision: https://phabricator.services.mozilla.com/D181165

UltraBlame original commit: b4e5d64ec56aeca640c11daebc87fa4e1aaf5135