Bug 1838879 - Centralize warning for nsXREDirProvider::mProfileDir. r=xpcom-reviewers,nika
nsXREDirProvider::mProfileDir will never be defined in a non-main process, so
adding a warning if we're missing one in that case just adds noise to the logs.

There was a lot of redundant ENSURE warning in code that ultimately depended
on mProfileDir, so this code rearranges things so that it all goes through
GetProfileStartupDir, then makes it so that we warn there if appropriate,
and do not warn anywhere else.

Looking back at the commit that added the comment about the
profile-after-change notification, this seems to be referring to a check on some
field mProfileNotified which no longer exists. Additionally, the component and
XPT registrations seem to be referring to NS_XPCOM_COMPONENT_REGISTRY_FILE and
NS_XPCOM_XPTI_REGISTRY_FILE, which also no longer exist. Therefore, I have
removed the comment and added a new one about the oddity of this particular
case.

Differential Revision: https://phabricator.services.mozilla.com/D181271

UltraBlame original commit: d2919829e4eb05d584b5611d26cc00dfb0ecc134