Bug 1689674 - Increase the profiler entry size for talos tests r=perftest-reviewers,sparky
Previously each test had their own profiler buffer entry size, and these buffer
sizes were really low. They were either 8MB or 16MB, which wasn't really useful
for normal profiling session.

This patch increases the profiler buffer size to 128MiB for all talos tests.
It's good to keep in mind that the whole 128MiB is not allocated at the start,
this is just a hard limit for the buffer, but the profiler allocates only the
amount of memory it needs. That's why in most of the profiling sessions this
limit will not be reached.

Also the second issue was that these buffer entry limits were only set for a
subset of the tests, and not all. That's why some of the tests were using only
1MiB which is the least amount of memory profiler can allocate currently. And
this is not enough for anything. I moved this limit to general defaults
section, so all the talos tests will benefit from this.

Differential Revision: https://phabricator.services.mozilla.com/D181092

UltraBlame original commit: 5f1f0f5559e239e38d683a443db7f8649327d102