Bug 1839178 - [devtools] Assert parseDeclaration offsets easier. r=devtools-reviewers,ochameau.
Seeing those offsets integer in the test case is a bit hard to assert that they
do actually refer to the expected part of the input.
This adds a `declarationText` property on the expected object that can be used
to properly asserts produced offsets.

Differential Revision: https://phabricator.services.mozilla.com/D181094

UltraBlame original commit: 46492ffbcaf882f70ee8e820e7ab92d9b96bd589