Bug 1838629 - Part 2: Move pre-alloc checks to allocator class r=sfink
This moves GCRuntime::checkAllocatorState to the CellAllocator class and
renames it to PreAllocChecks. This is conditionally compiled and is not present
in release builds. This simplifies the code a little.

GCRuntime::tryNewNurseryCell becomes static because it doesn't access |this|.

Differential Revision: https://phabricator.services.mozilla.com/D181190

UltraBlame original commit: 794cee44104764a2eae22f0a9e2b60d09352ced9