Bug 1519167 - Part 19: Implement abstract operations to retrieve Temporal.Calendar objects. r=spidermonkey-reviewers,mgaudet
Implements the following abstract operations:
- GetISO8601Calendar
- ToTemporalCalendar
- ToTemporalCalendarWithISODefault
- GetTemporalCalendarWithISODefault

Note: The `PlainTimeObject::getOrCreateCalendar()` is used to lazily create the
calendar object of a `Temporal.PlainTime` object. This method will likely be
removed when the spec is changed to represent built-in calendars as strings.

Differential Revision: https://phabricator.services.mozilla.com/D174275

UltraBlame original commit: 88f08c40096579bb9efbac74ef0c86844bd8af68