Bug 1519167 - Part 59: Implement Temporal.PlainTime.prototype.to{String,LocaleString,JSON}. r=spidermonkey-reviewers,sfink
The spec operation "RoundTime" rounds the mathematical value `quantity` to the
input increment. To implement this operation without loss of precision, we
can't use C++ `double` for the `quantity` variable, but instead need to
interpret the mathematical value, i.e. a real number ℝ, as a rational number ℚ.

For example when `unit="microseconds"`:

9.a Let quantity be nanosecond × 10^-3 + microsecond.
...
11. Let result be RoundNumberToIncrement(quantity, increment, roundingMode).

Where RoundNumberToIncrement is spec'ed as:
  1. Let quotient be x / increment.
  2-8. Let rounded be RoundOp(quotient).
  9. Return rounded × increment.

We compute `quantity` as the rational number `(nanosecond + microsecond × 10^3) / (10^3)`,
i.e. we scale it to nanoseconds. The quotient is then computed with
`x = (nanosecond + microsecond × 10^3) / (10^3)`:

    quotient
  = x / increment
  = ((nanosecond + microsecond × 10^3) / (10^3)) / increment
  = (nanosecond + microsecond × 10^3) / (increment × 10^3)

The same approach will be used in later patches when "RoundDuration" gets
implemented. The RoundDuration implementation has code comments explaining this
approach, that's why there's a comment in the RoundTime code which points to
RoundDuration.

Differential Revision: https://phabricator.services.mozilla.com/D174419

UltraBlame original commit: aadb81ba61bbcf38cb4d830b555f5e8134008a07